refactor: split extras_routes into one module per domain
extras_routes.py war mit 2339 Zeilen die zweitgroesste Datei im Backend und
hielt sechs fachlich unverwandte Router zusammen: Einstellungen, Saisons,
Widgets, DSGVO, PDF-Exporte und Auditlog. Wer an den Mail-Einstellungen
arbeitete, hatte die Zertifikatserzeugung im selben Puffer.

Die Abschnitte waren bereits sauber getrennt - eine Abhaengigkeitsanalyse fand
zwischen ihnen genau zwei Beruehrungspunkte. Entsprechend entstehen sechs
Module ohne Umbau der Logik:

  settings_routes.py   1150 Zeilen
  export_routes.py      556
  season_routes.py      372
  widget_routes.py      138
  dsgvo_routes.py       124
  audit_routes.py        30

Der zweite Beruehrungspunkt war `_safe_regex` - dieselben zwei Zeilen lagen
fuenfmal im Repository (extras, search, team, user, tournament). Statt sie ein
sechstes Mal zu kopieren, liegt die Funktion jetzt in
services/query_filters.py, und alle fuenf Kopien sind ersetzt. Damit ist auch
der entsprechende Punkt aus dem Audit erledigt.

Nachweis, dass sich nichts am Verhalten aendert: die Routentabellen vor und
nach dem Split wurden gegeneinander verglichen - 72 Routen, identisch in
Methode, Pfad und Handlernamen. Dazu 502 Backend-Tests gruen, Compile,
flake8-Gate, Secret-Scan und Doc-Links sauber.

Die Importe der neuen Module sind per pyflakes auf das tatsaechlich Benutzte
reduziert; Tests und `user_routes` verweisen jetzt direkt auf das jeweils
zustaendige Modul statt auf die Sammeldatei.
