--- backend/routes/extras_routes.py
@@ -1,2340 +1,0 @@
-"""Admin settings (email config, branding), Seasons/Circuits, Widgets, DSGVO, Audit Logs, PDF exports."""
-import os
-import re
-import secrets as secrets_lib
-from fastapi import APIRouter, HTTPException, Depends, Response
-from fastapi.responses import RedirectResponse, StreamingResponse
-from pydantic import BaseModel, EmailStr
-from typing import Optional, List, Literal
-from datetime import datetime, timezone
-import io
-import httpx
-
-from database import get_db
-from auth import require_admin, require_club_admin, require_role, require_super, get_current_user, get_optional_user
-from services.visibility import user_can_see
-from services.slug_utils import apply_slug_history, find_by_slug_or_history, slug_source_for_update, unique_slug
-from services.access_links import validate_access_link
-from services.competition_privacy import registration_match_snapshot
-from services.competition_read import load_competition_read_model, observe_structure_read
-from services.competition_standings import standings_for_structure
-from services.public_site_settings import PUBLIC_LEGAL_SOURCE_FIELDS, build_public_legal_settings
-from services.auth_settings import is_google_client_id, load_auth_settings
-from services.secret_store import encrypt_secret, secret_is_configured
-from models import now_utc, new_id
-from email_service import send_template, _get_email_config
-from pdf_service import (
-    pdf_participants, pdf_f1_leaderboard, pdf_matches, pdf_standings, pdf_checkin,
-    pdf_station_signs, pdf_qr_sign, pdf_certificate, pdf_certificates,
-)
-
-RESULT_EXPORT_STATUSES = {"completed", "results_published", "archived"}
-STAFF_EXPORT_ROLES = {"moderator", "tournament_admin", "club_admin", "superadmin"}
-
-# ---------- Settings ----------
-settings_router = APIRouter(prefix="/api/settings", tags=["settings"])
-
-DEFAULT_SOCIAL_LINKS = [
-    {"platform": "discord", "label": "Discord", "url": "https://discord.com/invite/thelionsquadesports", "enabled": True},
-    {"platform": "whatsapp", "label": "WhatsApp Kanal", "url": "https://whatsapp.com/channel/0029VaaWufTGU3BNG6VOxo1I", "enabled": True},
-    {"platform": "facebook", "label": "Facebook", "url": "https://www.facebook.com/thelionsquadesports", "enabled": True},
-    {"platform": "instagram", "label": "Instagram", "url": "https://instagram.com/thelionsquadesports", "enabled": True},
-    {"platform": "tiktok", "label": "TikTok", "url": "https://www.tiktok.com/@thelionsquadesports", "enabled": True},
-    {"platform": "youtube", "label": "YouTube", "url": "https://www.youtube.com/@TheLionSquadeSports", "enabled": True},
-    {"platform": "twitch", "label": "Twitch", "url": "https://www.twitch.tv/the_lion_squad_esports", "enabled": True},
-]
-
-SOCIAL_LEGACY_FIELDS = {
-    "discord": "discord_invite_url",
-    "whatsapp": "whatsapp_channel_url",
-    "facebook": "facebook_url",
-    "instagram": "instagram_url",
-    "tiktok": "tiktok_url",
-    "youtube": "youtube_url",
-}
-
-
-def _safe_regex(value: str | None, max_len: int = 80) -> str:
-    return re.escape((value or "").strip()[:max_len])
-
-
-class EmailSettings(BaseModel):
-    resend_api_key: Optional[str] = None
-    clear_resend_api_key: Optional[bool] = None
-    sender_name: Optional[str] = None
-    sender_email: Optional[str] = None
-    reply_to_email: Optional[str] = None
-    enabled: bool = True
-
-
-class SocialLinkSettings(BaseModel):
-    platform: Optional[str] = None
-    label: Optional[str] = None
-    url: Optional[str] = None
-    enabled: Optional[bool] = True
-
-
-class BrandingSettings(BaseModel):
-    club_name: Optional[str] = None
-    tagline: Optional[str] = None
-    site_title: Optional[str] = None
-    site_description: Optional[str] = None
-    primary_color: Optional[str] = None
-    logo_url: Optional[str] = None
-    logo_light_url: Optional[str] = None
-    logo_dark_url: Optional[str] = None
-    share_banner_url: Optional[str] = None
-    mascot_url: Optional[str] = None
-    qr_logo_url: Optional[str] = None
-    favicon_url: Optional[str] = None
-    favicon_light_url: Optional[str] = None
-    favicon_dark_url: Optional[str] = None
-    domain: Optional[str] = None
-    timezone: Optional[str] = None
-    contact_email: Optional[str] = None
-    imprint: Optional[str] = None
-    privacy_policy: Optional[str] = None
-    legal_name: Optional[str] = None
-    legal_form: Optional[str] = None
-    zvr_number: Optional[str] = None
-    street_address: Optional[str] = None
-    address_extra: Optional[str] = None
-    postal_code: Optional[str] = None
-    city: Optional[str] = None
-    state: Optional[str] = None
-    country: Optional[str] = None
-    registered_seat: Optional[str] = None
-    register_authority: Optional[str] = None
-    representative_name: Optional[str] = None
-    representative_role: Optional[str] = None
-    content_responsible: Optional[str] = None
-    phone: Optional[str] = None
-    privacy_contact_email: Optional[str] = None
-    hosting_provider: Optional[str] = None
-    hosting_country: Optional[str] = None
-    vat_number: Optional[str] = None
-    tournament_terms_url: Optional[str] = None
-    paid_tournaments_enabled: Optional[bool] = None
-    legal_extra: Optional[str] = None
-    privacy_extra: Optional[str] = None
-    terms_of_use: Optional[str] = None
-    discord_invite_url: Optional[str] = None
-    twitch_channel: Optional[str] = None
-    analytics_provider: Optional[Literal["", "google", "plausible"]] = None
-    google_analytics_id: Optional[str] = None
-    plausible_domain: Optional[str] = None
-    google_site_verification: Optional[str] = None
-    msvalidate_01: Optional[str] = None
-    indexnow_key: Optional[str] = None
-    whatsapp_channel_url: Optional[str] = None
-    # Social channels (Phase X — full social presence)
-    facebook_url: Optional[str] = None
-    instagram_url: Optional[str] = None
-    tiktok_url: Optional[str] = None
-    youtube_url: Optional[str] = None
-    social_links: Optional[List[SocialLinkSettings]] = None
-    # Phase E — Twitch Helix credentials
-    twitch_client_id: Optional[str] = None
-    twitch_client_secret: Optional[str] = None
-    clear_twitch_client_secret: Optional[bool] = None
-    twitch_live_detection: Optional[bool] = None
-    site_banner_enabled: Optional[bool] = None
-    site_banner_text: Optional[str] = None
-    site_banner_tone: Optional[Literal["info", "live", "warning", "success"]] = None
-    site_banner_mode: Optional[Literal["ticker", "static"]] = None
-    site_banner_speed_seconds: Optional[int] = None
-    site_banner_style: Optional[Literal["neon", "solid", "minimal"]] = None
-    site_banner_position: Optional[Literal["below_nav", "bottom_fixed", "above_footer"]] = None
-    site_banner_scope: Optional[Literal["all", "tournaments", "fastlap", "events", "news", "community", "servers", "members", "custom"]] = None
-    site_banner_path: Optional[str] = None
-    site_banner_audience: Optional[Literal["all", "logged_in", "members", "admins"]] = None
-    site_banner_link_url: Optional[str] = None
-    site_banner_link_label: Optional[str] = None
-    site_banner_starts_at: Optional[str] = None
-    site_banner_ends_at: Optional[str] = None
-
-
-class TestEmailBody(BaseModel):
-    to: EmailStr
-
-
-class NewsletterTriggerBody(BaseModel):
-    kind: Literal["news", "event"]
-    id: str
-    force: bool = False
-
-
-class IndexNowSubmitBody(BaseModel):
-    urls: Optional[List[str]] = None
-
-
-BannerTone = Literal["info", "live", "warning", "success"]
-BannerMode = Literal["ticker", "static"]
-BannerStyle = Literal["neon", "solid", "minimal"]
-BannerPosition = Literal["below_nav", "bottom_fixed", "above_footer"]
-BannerScope = Literal["all", "tournaments", "fastlap", "events", "news", "community", "servers", "members", "custom"]
-BannerAudience = Literal["all", "logged_in", "members", "admins"]
-BannerTemplate = Literal["custom", "live", "maintenance", "event", "registration", "discord"]
-
-
-class SiteBannerPayload(BaseModel):
-    title: Optional[str] = None
-    text: str
-    enabled: bool = True
-    priority: int = 50
-    tone: BannerTone = "info"
-    mode: BannerMode = "ticker"
-    speed_seconds: int = 22
-    style: BannerStyle = "neon"
-    position: BannerPosition = "below_nav"
-    scope: BannerScope = "all"
-    path: Optional[str] = None
-    audience: BannerAudience = "all"
-    link_url: Optional[str] = None
-    link_label: Optional[str] = None
-    starts_at: Optional[str] = None
-    ends_at: Optional[str] = None
-    template: BannerTemplate = "custom"
-
-
-class SiteBannerPatch(BaseModel):
-    title: Optional[str] = None
-    text: Optional[str] = None
-    enabled: Optional[bool] = None
-    priority: Optional[int] = None
-    tone: Optional[BannerTone] = None
-    mode: Optional[BannerMode] = None
-    speed_seconds: Optional[int] = None
-    style: Optional[BannerStyle] = None
-    position: Optional[BannerPosition] = None
-    scope: Optional[BannerScope] = None
-    path: Optional[str] = None
-    audience: Optional[BannerAudience] = None
-    link_url: Optional[str] = None
-    link_label: Optional[str] = None
-    starts_at: Optional[str] = None
-    ends_at: Optional[str] = None
-    template: Optional[BannerTemplate] = None
-
-
-class SiteBannerStatBody(BaseModel):
-    banner_id: str
-
-
-async def _newsletter_source(kind: str, source_id: str) -> dict:
-    db = get_db()
-    if kind == "news":
-        item = await db.news_posts.find_one({"$or": [{"id": source_id}, {"slug": source_id}]}, {"_id": 0})
-    elif kind == "event":
-        item = await db.events.find_one({"$or": [{"id": source_id}, {"slug": source_id}]}, {"_id": 0})
-    else:
-        item = None
-    if not item:
-        raise HTTPException(404, "Newsletter-Quelle nicht gefunden.")
-    return item
-
-
-class DiscordSettings(BaseModel):
-    webhook_url: Optional[str] = None
-    clear_webhook: Optional[bool] = None
-    username: Optional[str] = None
-    avatar_url: Optional[str] = None
-    enabled: bool = True
-
-
-class AuthSettings(BaseModel):
-    password_login_enabled: Optional[bool] = None
-    registration_enabled: Optional[bool] = None
-    google_login_enabled: Optional[bool] = None
-    google_registration_enabled: Optional[bool] = None
-    google_linking_enabled: Optional[bool] = None
-    google_client_id: Optional[str] = None
-
-
-SETTING_AUDIT_SECRET_FIELDS = {"resend_api_key", "smtp_pass", "webhook_url", "twitch_client_secret"}
-
-
-def _hide_branding_secrets(settings: dict) -> dict:
-    out = dict(settings or {})
-    if secret_is_configured(out.get("twitch_client_secret")):
-        out["twitch_client_secret_masked"] = "********"
-        out.pop("twitch_client_secret", None)
-    return out
-
-
-def _twitch_url(value: str | None) -> str:
-    raw = str(value or "").strip()
-    if not raw:
-        return "https://www.twitch.tv/the_lion_squad_esports"
-    if raw.startswith(("http://", "https://")):
-        return raw
-    return f"https://www.twitch.tv/{raw.lstrip('@')}"
-
-
-def _normalize_social_link(item: dict) -> dict | None:
-    platform = str(item.get("platform") or "").strip().lower()
-    label = str(item.get("label") or platform or "Link").strip()
-    url = str(item.get("url") or "").strip()
-    if not url:
-        return None
-    return {
-        "platform": platform or "custom",
-        "label": label,
-        "url": url,
-        "enabled": item.get("enabled") is not False,
-    }
-
-
-def _social_links_from_branding(settings: dict) -> list[dict]:
-    raw = settings.get("social_links")
-    if isinstance(raw, list) and raw:
-        links = [_normalize_social_link(item) for item in raw if isinstance(item, dict)]
-        return [link for link in links if link]
-    links = []
-    for default in DEFAULT_SOCIAL_LINKS:
-        item = dict(default)
-        field = SOCIAL_LEGACY_FIELDS.get(item["platform"])
-        if field:
-            item["url"] = settings.get(field) or item["url"]
-        elif item["platform"] == "twitch":
-            item["url"] = _twitch_url(settings.get("twitch_channel"))
-        normalized = _normalize_social_link(item)
-        if normalized:
-            links.append(normalized)
-    return links
-
-
-def _sync_legacy_social_fields(updates: dict) -> None:
-    raw = updates.get("social_links")
-    if not isinstance(raw, list):
-        return
-    normalized = [_normalize_social_link(item) for item in raw if isinstance(item, dict)]
-    updates["social_links"] = [item for item in normalized if item]
-    for item in updates["social_links"]:
-        platform = item.get("platform")
-        if platform in SOCIAL_LEGACY_FIELDS:
-            updates[SOCIAL_LEGACY_FIELDS[platform]] = item.get("url") or ""
-        elif platform == "twitch" and item.get("url"):
-            updates["twitch_channel"] = item["url"]
-
-
-def _normalize_setting_value(value):
-    if value is None:
-        return ""
-    if isinstance(value, str):
-        return value.strip()
-    if isinstance(value, list):
-        return [_normalize_setting_value(v) for v in value]
-    if isinstance(value, dict):
-        return {k: _normalize_setting_value(v) for k, v in sorted(value.items())}
-    return value
-
-
-def _changed_setting_fields(current: dict | None, updates: dict, unset: dict | None = None) -> list[str]:
-    current = current or {}
-    changed = set()
-    for key, value in updates.items():
-        if key in {"id", "updated_at"}:
-            continue
-        if key in SETTING_AUDIT_SECRET_FIELDS:
-            if value and _normalize_setting_value(current.get(key)) != _normalize_setting_value(value):
-                changed.add(key)
-            continue
-        if _normalize_setting_value(current.get(key)) != _normalize_setting_value(value):
-            changed.add(key)
-    for key in (unset or {}):
-        if _normalize_setting_value(current.get(key)) != "":
-            changed.add(key)
-    return sorted(changed)
-
-
-async def _audit_settings_change(db, action: str, setting_id: str, actor_id: str, changed_fields: list[str]) -> None:
-    if not changed_fields:
-        return
-    await db.audit_logs.insert_one({
-        "id": new_id(),
-        "action": action,
-        "target_id": setting_id,
-        "actor_id": actor_id,
-        "data": {"changed_fields": changed_fields},
-        "created_at": now_utc().isoformat(),
-    })
-
-
-def _parse_public_dt(value: str | None):
-    if not value:
-        return None
-    try:
-        dt = datetime.fromisoformat(str(value).replace("Z", "+00:00"))
-        if dt.tzinfo is None:
-            dt = dt.replace(tzinfo=timezone.utc)
-        return dt
-    except ValueError:
-        return None
-
-
-def _normalize_banner_link(value: str | None) -> str:
-    link_url = str(value or "").strip()
-    if link_url and not link_url.startswith(("http://", "https://", "/")):
-        link_url = "https://" + link_url
-    return link_url
-
-
-def _banner_active(doc: dict, user: dict | None) -> bool:
-    if not doc.get("enabled", True):
-        return False
-    if not str(doc.get("text") or "").strip():
-        return False
-    now = now_utc()
-    starts_at = _parse_public_dt(doc.get("starts_at"))
-    ends_at = _parse_public_dt(doc.get("ends_at"))
-    if starts_at and starts_at > now:
-        return False
-    if ends_at and ends_at < now:
-        return False
-    audience = doc.get("audience") or "all"
-    admin_roles = {"moderator", "tournament_admin", "club_admin", "superadmin"}
-    if audience == "all":
-        return True
-    if audience == "logged_in":
-        return bool(user)
-    if audience == "members":
-        return bool(user and (user.get("is_club_member") or user.get("role") in admin_roles))
-    if audience == "admins":
-        return bool(user and user.get("role") in admin_roles)
-    return False
-
-
-def _public_banner_doc(doc: dict, stats: dict | None = None) -> dict:
-    return {
-        "id": doc.get("id"),
-        "enabled": bool(doc.get("enabled", True)),
-        "title": doc.get("title") or "",
-        "text": str(doc.get("text") or "").strip(),
-        "tone": doc.get("tone") or "info",
-        "mode": doc.get("mode") or "ticker",
-        "speed_seconds": max(8, min(180, int(doc.get("speed_seconds") or 22))),
-        "style": doc.get("style") or "neon",
-        "position": doc.get("position") or "below_nav",
-        "scope": doc.get("scope") or "all",
-        "path": str(doc.get("path") or "").strip(),
-        "audience": doc.get("audience") or "all",
-        "link_url": _normalize_banner_link(doc.get("link_url")),
-        "link_label": str(doc.get("link_label") or "").strip(),
-        "priority": int(doc.get("priority") or 0),
-        "template": doc.get("template") or "custom",
-        "source": doc.get("source") or "manual",
-        "stats": {
-            "impressions": int((stats or {}).get("impressions") or doc.get("impressions") or 0),
-            "clicks": int((stats or {}).get("clicks") or doc.get("clicks") or 0),
-        },
-    }
-
-
-def _parse_dt_any(value):
-    if not value:
-        return None
-    if isinstance(value, datetime):
-        return value if value.tzinfo else value.replace(tzinfo=timezone.utc)
-    return _parse_public_dt(str(value))
-
-
-async def _auto_site_banners(db) -> list[dict]:
-    now = now_utc()
-    items: list[dict] = []
-    servers = await db.game_servers.find(
-        {"status": "maintenance", "site_banner_enabled": True, "is_active": {"$ne": False}},
-        {"_id": 0, "id": 1, "name": 1, "maintenance_note": 1, "maintenance_until": 1, "visibility": 1},
-    ).sort("sort_order", 1).to_list(3)
-    for server in servers:
-        until = server.get("maintenance_until")
-        text = f"{server.get('name') or 'Server'} ist in Wartung"
-        if server.get("maintenance_note"):
-            text += f": {server['maintenance_note']}"
-        if until:
-            text += f" · bis {until}"
-        items.append({
-            "id": f"auto-server-maintenance-{server.get('id')}",
-            "source": "auto",
-            "template": "maintenance",
-            "enabled": True,
-            "priority": 80,
-            "text": text,
-            "tone": "warning",
-            "mode": "ticker",
-            "speed_seconds": 24,
-            "style": "minimal",
-            "position": "below_nav",
-            "scope": "servers",
-            "audience": "members" if server.get("visibility") == "members" else "logged_in" if server.get("visibility") == "community" else "all",
-            "link_url": "/servers",
-            "link_label": "Server ansehen",
-        })
-
-    tournaments = await db.tournaments.find(
-        {"status": {"$in": ["registration_open", "check_in", "scheduled", "registration_closed"]}, "site_banner_enabled": True, "is_public": {"$ne": False}},
-        {"_id": 0, "id": 1, "slug": 1, "title": 1, "status": 1, "start_date": 1},
-    ).sort("start_date", 1).to_list(10)
-    for t in tournaments:
-        start_dt = _parse_dt_any(t.get("start_date"))
-        if t.get("status") == "check_in":
-            text = f"Check-in offen: {t.get('title') or 'Turnier'}"
-            template, tone, priority = "registration", "warning", 85
-        elif t.get("status") == "registration_open":
-            text = f"Anmeldung offen: {t.get('title') or 'Turnier'}"
-            template, tone, priority = "registration", "success", 75
-        elif start_dt and now <= start_dt and (start_dt - now).total_seconds() <= 24 * 3600:
-            text = f"Turnier startet bald: {t.get('title') or 'Turnier'}"
-            template, tone, priority = "event", "info", 70
-        else:
-            continue
-        items.append({
-            "id": f"auto-tournament-{template}-{t.get('id')}",
-            "source": "auto",
-            "template": template,
-            "enabled": True,
-            "priority": priority,
-            "text": text,
-            "tone": tone,
-            "mode": "ticker",
-            "speed_seconds": 24,
-            "style": "neon",
-            "position": "below_nav",
-            "scope": "tournaments",
-            "audience": "all",
-            "link_url": f"/tournaments/{t.get('slug') or t.get('id')}",
-            "link_label": "Zum Turnier",
-        })
-
-    fastlaps = await db.f1_challenges.find(
-        {"status": {"$in": ["registration_open", "check_in", "scheduled", "registration_closed"]}, "site_banner_enabled": True, "visibility": {"$ne": "internal"}},
-        {"_id": 0, "id": 1, "slug": 1, "title": 1, "status": 1, "start_date": 1},
-    ).sort("start_date", 1).to_list(10)
-    for challenge in fastlaps:
-        start_dt = _parse_dt_any(challenge.get("start_date"))
-        if challenge.get("status") == "registration_open":
-            text = f"Fast-Lap Einreichung offen: {challenge.get('title') or 'Challenge'}"
-            template, tone, priority = "registration", "success", 75
-        elif start_dt and now <= start_dt and (start_dt - now).total_seconds() <= 24 * 3600:
-            text = f"Fast-Lap startet bald: {challenge.get('title') or 'Challenge'}"
-            template, tone, priority = "event", "info", 70
-        else:
-            continue
-        items.append({
-            "id": f"auto-fastlap-{template}-{challenge.get('id')}",
-            "source": "auto",
-            "template": template,
-            "enabled": True,
-            "priority": priority,
-            "text": text,
-            "tone": tone,
-            "mode": "ticker",
-            "speed_seconds": 24,
-            "style": "neon",
-            "position": "below_nav",
-            "scope": "fastlap",
-            "audience": "all",
-            "link_url": f"/fastlap/{challenge.get('slug') or challenge.get('id')}",
-            "link_label": "Zur Challenge",
-        })
-    return items
-
-
-@settings_router.get("/public")
-async def public_settings(response: Response):
-    """Public-safe settings for branding on public pages."""
-    response.headers["Cache-Control"] = "no-store"
-    db = get_db()
-    b = await db.settings.find_one({"id": "branding"}) or {}
-    b.pop("_id", None)
-    domain = (b.get("domain") or "https://lionsquad.at").strip()
-    if domain and not domain.startswith(("http://", "https://")):
-        domain = "https://" + domain
-    tagline = b.get("tagline", "eSports Verein")
-    if str(tagline).strip().lower() == "esports arena":
-        tagline = "eSports Verein"
-    legal_settings = build_public_legal_settings(b)
-    legal_settings.pop("contact_ready", None)
-    legal_settings.pop("missing_legal_fields", None)
-    return {
-        "club_name": b.get("club_name", "THE LION SQUAD"),
-        "tagline": tagline,
-        "site_title": b.get("site_title") or "THE LION SQUAD - eSPORTS",
-        "site_description": b.get("site_description") or "Gaming und eSports Verein aus Tirol mit Community, Turnieren, Fast-Lap-Challenges, Events, Mitgliedschaft und Vereinsleben.",
-        "primary_color": b.get("primary_color", "#29B6E8"),
-        "logo_url": b.get("logo_url"),
-        "logo_light_url": b.get("logo_light_url"),
-        "logo_dark_url": b.get("logo_dark_url"),
-        "share_banner_url": b.get("share_banner_url"),
-        "mascot_url": b.get("mascot_url"),
-        "qr_logo_url": b.get("qr_logo_url"),
-        "favicon_url": b.get("favicon_url"),
-        "favicon_light_url": b.get("favicon_light_url"),
-        "favicon_dark_url": b.get("favicon_dark_url"),
-        "domain": domain,
-        "timezone": b.get("timezone") or "Europe/Vienna",
-        **legal_settings,
-        "discord_invite_url": b.get("discord_invite_url") or "https://discord.com/invite/thelionsquadesports",
-        "twitch_channel": b.get("twitch_channel") or "the_lion_squad_esports",
-        "whatsapp_channel_url": b.get("whatsapp_channel_url") or "https://whatsapp.com/channel/0029VaaWufTGU3BNG6VOxo1I",
-        "analytics_provider": b.get("analytics_provider") or "",
-        "google_analytics_id": b.get("google_analytics_id") or "",
-        "plausible_domain": b.get("plausible_domain") or "",
-        "google_site_verification": b.get("google_site_verification") or "",
-        "msvalidate_01": b.get("msvalidate_01") or "",
-        "indexnow_key": b.get("indexnow_key") or "",
-        "facebook_url": b.get("facebook_url") or "https://www.facebook.com/thelionsquadesports",
-        "instagram_url": b.get("instagram_url") or "https://instagram.com/thelionsquadesports",
-        "tiktok_url": b.get("tiktok_url") or "https://www.tiktok.com/@thelionsquadesports",
-        "youtube_url": b.get("youtube_url") or "https://www.youtube.com/@TheLionSquadeSports",
-        "social_links": _social_links_from_branding(b),
-        **(await load_auth_settings(db)),
-    }
-
-
-@settings_router.get("/site-banner")
-async def get_site_banner(response: Response, me: dict | None = Depends(get_optional_user)):
-    response.headers["Cache-Control"] = "no-store"
-    data = await list_site_banners(response, me)
-    first = (data.get("items") or [None])[0]
-    if not first:
-        return {"enabled": False}
-    return first
-
-
-@settings_router.get("/site-banners")
-async def list_site_banners(response: Response, me: dict | None = Depends(get_optional_user)):
-    response.headers["Cache-Control"] = "no-store"
-    db = get_db()
-    manual = await db.site_banners.find({}, {"_id": 0}).sort([("priority", -1), ("updated_at", -1)]).to_list(100)
-    # The legacy branding banner is intentionally no longer emitted. The
-    # Banner-Manager owns public notice bars from here on.
-    all_docs = manual + await _auto_site_banners(db)
-    active = [doc for doc in all_docs if _banner_active(doc, me)]
-    stats_rows = await db.site_banner_stats.find({"id": {"$in": [doc["id"] for doc in active if doc.get("id")]}}, {"_id": 0}).to_list(200)
-    stats = {row["id"]: row for row in stats_rows}
-    items = [_public_banner_doc(doc, stats.get(doc.get("id"))) for doc in active]
-    items.sort(key=lambda row: (int(row.get("priority") or 0), row.get("id") or ""), reverse=True)
-    return {"items": items[:8]}
-
-
-@settings_router.get("/site-banners/admin")
-async def admin_site_banners(me: dict = Depends(require_admin())):
-    db = get_db()
-    rows = await db.site_banners.find({}, {"_id": 0}).sort([("priority", -1), ("updated_at", -1)]).to_list(200)
-    stats_rows = await db.site_banner_stats.find({"id": {"$in": [row["id"] for row in rows if row.get("id")]}}, {"_id": 0}).to_list(200)
-    stats = {row["id"]: row for row in stats_rows}
-    return [_public_banner_doc(row, stats.get(row.get("id"))) | {
-        "starts_at": row.get("starts_at") or "",
-        "ends_at": row.get("ends_at") or "",
-    } for row in rows]
-
-
-@settings_router.post("/site-banners/admin")
-async def create_site_banner(body: SiteBannerPayload, me: dict = Depends(require_admin())):
-    db = get_db()
-    now = now_utc().isoformat()
-    doc = {
-        **body.model_dump(),
-        "id": new_id(),
-        "created_at": now,
-        "updated_at": now,
-        "created_by": me["id"],
-        "source": "manual",
-    }
-    await db.site_banners.insert_one(doc)
-    doc.pop("_id", None)
-    return doc
-
-
-@settings_router.patch("/site-banners/admin/{banner_id}")
-@settings_router.put("/site-banners/admin/{banner_id}")
-async def update_site_banner(banner_id: str, body: SiteBannerPatch, me: dict = Depends(require_admin())):
-    db = get_db()
-    updates = {k: v for k, v in body.model_dump(exclude_unset=True).items() if v is not None}
-    if not updates:
-        raise HTTPException(400, "Keine Änderungen.")
-    updates["updated_at"] = now_utc().isoformat()
-    res = await db.site_banners.update_one({"id": banner_id}, {"$set": updates})
-    if res.matched_count == 0:
-        raise HTTPException(404, "Banner nicht gefunden.")
-    return await db.site_banners.find_one({"id": banner_id}, {"_id": 0})
-
-
-@settings_router.delete("/site-banners/admin/{banner_id}")
-async def delete_site_banner(banner_id: str, me: dict = Depends(require_admin())):
-    db = get_db()
-    res = await db.site_banners.delete_one({"id": banner_id})
-    if res.deleted_count == 0:
-        raise HTTPException(404, "Banner nicht gefunden.")
-    return {"ok": True}
-
-
-@settings_router.post("/site-banners/impression")
-async def track_site_banner_impression(body: SiteBannerStatBody):
-    db = get_db()
-    now = now_utc().isoformat()
-    await db.site_banner_stats.update_one(
-        {"id": body.banner_id},
-        {"$inc": {"impressions": 1}, "$set": {"last_impression_at": now}, "$setOnInsert": {"id": body.banner_id}},
-        upsert=True,
-    )
-    return {"ok": True}
-
-
-@settings_router.post("/site-banners/click")
-async def track_site_banner_click(body: SiteBannerStatBody):
-    db = get_db()
-    now = now_utc().isoformat()
-    await db.site_banner_stats.update_one(
-        {"id": body.banner_id},
-        {"$inc": {"clicks": 1}, "$set": {"last_click_at": now}, "$setOnInsert": {"id": body.banner_id}},
-        upsert=True,
-    )
-    return {"ok": True}
-
-
-@settings_router.get("/email")
-async def get_email_settings(me: dict = Depends(require_club_admin())):
-    db = get_db()
-    s = await db.settings.find_one({"id": "email"}, {"_id": 0}) or {}
-    # Mask the API key
-    if s.get("resend_api_key"):
-        s["resend_api_key_masked"] = "********"
-        s.pop("resend_api_key", None)
-    return s
-
-
-@settings_router.put("/email")
-async def update_email_settings(body: EmailSettings, me: dict = Depends(require_club_admin())):
-    db = get_db()
-    updates = {k: v for k, v in body.model_dump(exclude_unset=True).items() if v is not None}
-    clear_secret = bool(updates.pop("clear_resend_api_key", False))
-    unset = {"resend_api_key": ""} if clear_secret else {}
-    # Only overwrite api_key if a non-empty value given
-    if "resend_api_key" in updates and not updates["resend_api_key"]:
-        updates.pop("resend_api_key")
-    elif "resend_api_key" in updates:
-        updates["resend_api_key"] = encrypt_secret(updates["resend_api_key"])
-    current = await db.settings.find_one({"id": "email"}, {"_id": 0}) or {}
-    changed_fields = _changed_setting_fields(current, updates, unset)
-    if not changed_fields:
-        return {"ok": True, "changed": False}
-    updates["updated_at"] = now_utc().isoformat()
-    operation = {"$set": updates, "$setOnInsert": {"id": "email"}}
-    if unset:
-        operation["$unset"] = unset
-    await db.settings.update_one({"id": "email"}, operation, upsert=True)
-    await _audit_settings_change(db, "settings.email.update", "email", me["id"], changed_fields)
-    return {"ok": True, "changed": True}
-
-
-@settings_router.post("/email/test")
-async def send_test(body: TestEmailBody, me: dict = Depends(require_club_admin())):
-    res = await send_template("test", body.to, branding="THE LION SQUAD", queue=False)
-    return res
-
-
-@settings_router.post("/newsletter/preview")
-async def newsletter_preview(body: NewsletterTriggerBody, me: dict = Depends(require_admin())):
-    from services.notification_preferences import newsletter_recipients
-    item = await _newsletter_source(body.kind, body.id)
-    visibility = item.get("visibility") or "public"
-    recipients = await newsletter_recipients(visibility)
-    return {
-        "kind": body.kind,
-        "source_id": item.get("id"),
-        "title": item.get("title") or item.get("name"),
-        "visibility": visibility,
-        "already_sent_at": item.get("newsletter_sent_at"),
-        "already_sent_count": item.get("newsletter_sent_count") or 0,
-        "recipients": len(recipients),
-        "sample": [
-            {
-                "id": u.get("id"),
-                "email": u.get("email"),
-                "display_name": u.get("display_name") or u.get("username"),
-            }
-            for u in recipients[:10]
-        ],
-    }
-
-
-@settings_router.post("/newsletter/send")
-async def newsletter_send(body: NewsletterTriggerBody, me: dict = Depends(require_admin())):
-    db = get_db()
-    from services.notification_preferences import enqueue_newsletter_for_item
-    item = await _newsletter_source(body.kind, body.id)
-    if item.get("newsletter_sent_at") and not body.force:
-        return {
-            "ok": True,
-            "skipped": True,
-            "reason": "already_sent",
-            "queued": 0,
-            "sent_at": item.get("newsletter_sent_at"),
-            "sent_count": item.get("newsletter_sent_count") or 0,
-        }
-    suffix = f":manual:{now_utc().isoformat()}" if body.force else ""
-    result = await enqueue_newsletter_for_item(body.kind, item, dedupe_suffix=suffix)
-    queued = int(result.get("queued") or 0)
-    collection = db.news_posts if body.kind == "news" else db.events
-    await collection.update_one(
-        {"id": item.get("id")},
-        {"$set": {
-            "newsletter_sent_at": now_utc().isoformat(),
-            "newsletter_sent_count": queued,
-            "newsletter_sent_by": me["id"],
-        }},
-    )
-    await db.audit_logs.insert_one({
-        "id": new_id(),
-        "action": "newsletter.manual_send",
-        "actor_id": me["id"],
-        "target_id": item.get("id"),
-        "data": {"kind": body.kind, "queued": queued, "force": body.force},
-        "created_at": now_utc().isoformat(),
-    })
-    return {"ok": True, **result}
-
-
-# ---------- Phase 8: SMTP & Mail Queue ----------
-class SmtpSettings(BaseModel):
-    provider: Optional[Literal["smtp", "resend"]] = None
-    smtp_host: Optional[str] = None
-    smtp_port: Optional[int] = None
-    smtp_user: Optional[str] = None
-    smtp_pass: Optional[str] = None
-    clear_smtp_pass: Optional[bool] = None
-    smtp_auth: Optional[Literal["auto", "login", "none"]] = None
-    smtp_security: Optional[Literal["auto", "starttls", "tls", "none"]] = None
-    smtp_tls_verify: Optional[bool] = None
-    smtp_envelope_from: Optional[str] = None
-    smtp_helo_name: Optional[str] = None
-    sender_name: Optional[str] = None
-    sender_email: Optional[str] = None
-    reply_to_email: Optional[str] = None
-    message_id_domain: Optional[str] = None
-    enabled: Optional[bool] = None
-
-
-@settings_router.get("/smtp")
-async def get_smtp_settings(me: dict = Depends(require_club_admin())):
-    db = get_db()
-    s = await db.settings.find_one({"id": "mail"}, {"_id": 0}) or {}
-    if s.get("smtp_pass"):
-        s["smtp_pass_masked"] = "••••••••"
-    s.pop("smtp_pass", None)
-    s.setdefault("provider", "smtp" if s.get("smtp_host") else "resend")
-    s.setdefault("smtp_auth", "login")
-    if s.get("smtp_auth") == "auto":
-        s["smtp_auth"] = "login"
-    s.setdefault("smtp_security", "auto")
-    s.setdefault("smtp_port", 587)
-    s.setdefault("smtp_tls_verify", False)
-    s.setdefault("smtp_helo_name", "")
-    s.setdefault("enabled", True)
-    s.setdefault("reply_to_email", s.get("sender_email") or "")
-    s.setdefault("message_id_domain", "")
-    return s
-
-
-@settings_router.put("/smtp")
-async def update_smtp_settings(body: SmtpSettings, me: dict = Depends(require_club_admin())):
-    db = get_db()
-    updates = {k: v for k, v in body.model_dump(exclude_unset=True).items() if v is not None}
-    clear_secret = bool(updates.pop("clear_smtp_pass", False))
-    unset = {"smtp_pass": ""} if clear_secret else {}
-    # Only overwrite password if a non-empty value given
-    if "smtp_pass" in updates and not updates["smtp_pass"]:
-        updates.pop("smtp_pass")
-    elif "smtp_pass" in updates:
-        updates["smtp_pass"] = encrypt_secret(updates["smtp_pass"])
-    current = await db.settings.find_one({"id": "mail"}, {"_id": 0}) or {}
-    changed_fields = _changed_setting_fields(current, updates, unset)
-    if not changed_fields:
-        return {"ok": True, "changed": False}
-    updates["updated_at"] = now_utc().isoformat()
-    operation = {"$set": updates, "$setOnInsert": {"id": "mail"}}
-    if unset:
-        operation["$unset"] = unset
-    await db.settings.update_one({"id": "mail"}, operation, upsert=True)
-    await _audit_settings_change(db, "settings.smtp.update", "mail", me["id"], changed_fields)
-    return {"ok": True, "changed": True}
-
-
-@settings_router.post("/smtp/test")
-async def smtp_send_test(body: TestEmailBody, me: dict = Depends(require_club_admin())):
-    from services.mail_queue import smtp_test
-    return await smtp_test(body.to)
-
-
-@settings_router.post("/smtp/diagnose")
-async def smtp_diagnose(body: TestEmailBody, me: dict = Depends(require_club_admin())):
-    from services.mail_queue import smtp_diagnose as run_smtp_diagnose
-    return await run_smtp_diagnose(body.to)
-
-
-@settings_router.get("/smtp/deliverability")
-async def smtp_deliverability(me: dict = Depends(require_club_admin())):
-    from services.mail_queue import smtp_deliverability as run_smtp_deliverability
-    return await run_smtp_deliverability()
-
-
-@settings_router.get("/mail-queue")
-async def list_mail_queue(status: Optional[str] = None, limit: int = 100,
-                          me: dict = Depends(require_club_admin())):
-    db = get_db()
-    q = {}
-    if status:
-        q["status"] = status
-    safe_limit = max(1, min(int(limit or 100), 500))
-    jobs = await db.mail_jobs.find(q, {"_id": 0, "html": 0}).sort("created_at", -1).to_list(safe_limit)
-    return jobs
-
-
-@settings_router.get("/mail-queue/stats")
-async def mail_queue_statistics(me: dict = Depends(require_club_admin())):
-    from services.mail_queue import mail_queue_stats
-    return await mail_queue_stats()
-
-
-@settings_router.post("/mail-queue/process")
-async def process_queue_now(me: dict = Depends(require_club_admin())):
-    from services.mail_queue import process_mail_queue
-    return await process_mail_queue(batch=20)
-
-
-@settings_router.post("/mail-queue/recover")
-async def recover_mail_queue(me: dict = Depends(require_club_admin())):
-    from services.mail_queue import recover_stale_sending_jobs
-    return {"recovered": await recover_stale_sending_jobs()}
-
-
-@settings_router.post("/mail-queue/retry-failed")
-async def retry_failed_mail_jobs(me: dict = Depends(require_club_admin())):
-    from services.mail_queue import retry_failed_jobs
-    return {"queued": await retry_failed_jobs()}
-
-
-@settings_router.delete("/mail-queue/cleanup")
-async def cleanup_mail_queue(days: int = 30, me: dict = Depends(require_club_admin())):
-    from services.mail_queue import cleanup_sent_jobs
-    return {"deleted": await cleanup_sent_jobs(days=days)}
-
-
-@settings_router.post("/mail-queue/{job_id}/retry")
-async def retry_mail_job(job_id: str, me: dict = Depends(require_club_admin())):
-    db = get_db()
-    res = await db.mail_jobs.update_one(
-        {"id": job_id},
-        {"$set": {
-            "status": "pending",
-            "attempts": 0,
-            "next_attempt_at": now_utc().isoformat(),
-            "last_error": None,
-            "updated_at": now_utc().isoformat(),
-        }},
-    )
-    if res.matched_count == 0:
-        raise HTTPException(404, "Job nicht gefunden")
-    return {"ok": True}
-
-
-@settings_router.delete("/mail-queue/{job_id}")
-async def delete_mail_job(job_id: str, me: dict = Depends(require_club_admin())):
-    db = get_db()
-    res = await db.mail_jobs.delete_one({"id": job_id})
-    if res.deleted_count == 0:
-        raise HTTPException(404, "Job nicht gefunden")
-    return {"ok": True}
-
-
-@settings_router.get("/branding")
-async def get_branding(response: Response, me: dict = Depends(require_club_admin())):
-    """Returns the branding doc, merging in social-default URLs so the admin form
-    pre-fills with sensible defaults instead of empty fields."""
-    response.headers["Cache-Control"] = "no-store"
-    db = get_db()
-    saved = (await db.settings.find_one({"id": "branding"}, {"_id": 0})) or {}
-    defaults = {
-        "site_title": "THE LION SQUAD - eSPORTS",
-        "discord_invite_url": "https://discord.com/invite/thelionsquadesports",
-        "twitch_channel": "the_lion_squad_esports",
-        "whatsapp_channel_url": "https://whatsapp.com/channel/0029VaaWufTGU3BNG6VOxo1I",
-        "facebook_url": "https://www.facebook.com/thelionsquadesports",
-        "instagram_url": "https://instagram.com/thelionsquadesports",
-        "tiktok_url": "https://www.tiktok.com/@thelionsquadesports",
-        "youtube_url": "https://www.youtube.com/@TheLionSquadeSports",
-    }
-    for k, v in defaults.items():
-        if not saved.get(k):
-            saved[k] = v
-    saved["social_links"] = _social_links_from_branding(saved)
-    return _hide_branding_secrets(saved)
-
-
-@settings_router.put("/branding")
-async def update_branding(body: BrandingSettings, me: dict = Depends(require_club_admin())):
-    db = get_db()
-    nullable_fields = set(BrandingSettings.model_fields.keys())
-    raw = body.model_dump(exclude_unset=True)
-    clear_twitch_secret = bool(raw.pop("clear_twitch_client_secret", False))
-    unset = {"twitch_client_secret": ""} if clear_twitch_secret else {}
-    updates = {k: v for k, v in raw.items() if v is not None or k in nullable_fields}
-    if updates.get("twitch_client_secret"):
-        updates["twitch_client_secret"] = encrypt_secret(updates["twitch_client_secret"])
-    _sync_legacy_social_fields(updates)
-    current = await db.settings.find_one({"id": "branding"}, {"_id": 0}) or {}
-    changed_fields = _changed_setting_fields(current, updates, unset)
-    if not changed_fields:
-        return current or {"ok": True, "changed": False}
-    updates["updated_at"] = now_utc().isoformat()
-    if set(changed_fields) & PUBLIC_LEGAL_SOURCE_FIELDS:
-        updates["legal_updated_at"] = updates["updated_at"]
-    operation = {"$set": updates, "$setOnInsert": {"id": "branding"}}
-    if unset:
-        operation["$unset"] = unset
-    await db.settings.update_one({"id": "branding"}, operation, upsert=True)
-    if clear_twitch_secret:
-        await db.settings.delete_one({"id": "twitch_app_token"})
-    await _audit_settings_change(db, "settings.branding.update", "branding", me["id"], changed_fields)
-    saved = await db.settings.find_one({"id": "branding"}, {"_id": 0})
-    return _hide_branding_secrets(saved) if saved else {"ok": True}
-
-
-@settings_router.get("/auth")
-async def get_auth_settings(response: Response, me: dict = Depends(require_super())):
-    """Central login & Google configuration for the admin area."""
-    response.headers["Cache-Control"] = "no-store"
-    db = get_db()
-    return await load_auth_settings(db)
-
-
-@settings_router.put("/auth")
-async def update_auth_settings(body: AuthSettings, me: dict = Depends(require_super())):
-    db = get_db()
-    raw = body.model_dump(exclude_unset=True)
-    updates = {k: v for k, v in raw.items() if isinstance(v, bool)}
-    if "google_client_id" in raw:
-        client_id = str(raw.get("google_client_id") or "").strip()
-        if client_id and not is_google_client_id(client_id):
-            raise HTTPException(400, "Ungültige Google Web Client ID.")
-        updates["google_client_id"] = client_id
-        if not client_id:
-            updates.update({
-                "google_login_enabled": False,
-                "google_registration_enabled": False,
-                "google_linking_enabled": False,
-            })
-    if not updates:
-        return await load_auth_settings(db)
-    current = await db.settings.find_one({"id": "auth"}, {"_id": 0}) or {}
-    changed_fields = _changed_setting_fields(current, updates)
-    if not changed_fields:
-        return await load_auth_settings(db)
-    updates["updated_at"] = now_utc().isoformat()
-    await db.settings.update_one(
-        {"id": "auth"}, {"$set": updates, "$setOnInsert": {"id": "auth"}}, upsert=True,
-    )
-    await _audit_settings_change(db, "settings.auth.update", "auth", me["id"], changed_fields)
-    return await load_auth_settings(db)
-
-
-@settings_router.post("/auth/google/test")
-async def test_google_auth_settings(me: dict = Depends(require_super())):
-    """Validate the stored client shape and Google's discovery availability."""
-    db = get_db()
-    settings = await load_auth_settings(db)
-    if not settings["google_configured"]:
-        raise HTTPException(400, "Bitte zuerst eine gültige Google Web Client ID speichern.")
-    provider_available = False
-    try:
-        async with httpx.AsyncClient(timeout=5) as client:
-            response = await client.get("https://accounts.google.com/.well-known/openid-configuration")
-            provider_available = response.status_code == 200 and bool(response.json().get("jwks_uri"))
-    except (httpx.HTTPError, ValueError):
-        provider_available = False
-    return {
-        "ok": provider_available,
-        "configured": True,
-        "provider_available": provider_available,
-        "expected_origin": os.environ.get("FRONTEND_URL", "").rstrip("/"),
-        "message": "Google ist erreichbar. Bitte den Anmeldebutton zusätzlich mit einem Testkonto prüfen." if provider_available else "Google ist derzeit nicht erreichbar.",
-    }
-
-
-@settings_router.post("/indexnow/submit")
-async def submit_indexnow(body: IndexNowSubmitBody, me: dict = Depends(require_club_admin())):
-    db = get_db()
-    branding = await db.settings.find_one({"id": "branding"}, {"_id": 0}) or {}
-    key = (branding.get("indexnow_key") or "").strip()
-    if not key:
-        raise HTTPException(400, "IndexNow-Key fehlt. Bitte im Branding hinterlegen.")
-    domain = (branding.get("domain") or "https://lionsquad.at").strip().rstrip("/")
-    if not domain.startswith(("http://", "https://")):
-        domain = "https://" + domain
-    host = domain.replace("https://", "").replace("http://", "").split("/")[0]
-    urls = body.urls or [domain, f"{domain}/sitemap.xml"]
-    urls = [u if str(u).startswith(("http://", "https://")) else f"{domain}/{str(u).lstrip('/')}" for u in urls]
-    payload = {
-        "host": host,
-        "key": key,
-        "keyLocation": f"{domain}/indexnow-key.txt",
-        "urlList": urls[:100],
-    }
-    async with httpx.AsyncClient(timeout=10) as client:
-        r = await client.post("https://api.indexnow.org/indexnow", json=payload)
-    if r.status_code >= 400:
-        raise HTTPException(r.status_code, f"IndexNow fehlgeschlagen: {r.text[:300]}")
-    return {"ok": True, "submitted": len(payload["urlList"]), "status": r.status_code}
-
-
-@settings_router.get("/indexnow-key.txt")
-async def indexnow_key_file():
-    db = get_db()
-    branding = await db.settings.find_one({"id": "branding"}, {"_id": 0}) or {}
-    key = (branding.get("indexnow_key") or "").strip()
-    if not key:
-        raise HTTPException(404, "IndexNow-Key nicht gesetzt")
-    return Response(content=key, media_type="text/plain")
-
-
-@settings_router.get("/email/logs")
-async def email_logs(me: dict = Depends(require_club_admin())):
-    db = get_db()
-    return await db.email_logs.find({}, {"_id": 0}).sort("created_at", -1).to_list(200)
-
-
-# ---- Discord webhook ----
-@settings_router.get("/discord")
-async def get_discord(me: dict = Depends(require_club_admin())):
-    db = get_db()
-    s = await db.settings.find_one({"id": "discord"}, {"_id": 0}) or {}
-    s["configured"] = bool(s.get("webhook_url"))
-    if s.get("webhook_url"):
-        s["webhook_url_masked"] = "https://discord.com/api/webhooks/…"
-        s.pop("webhook_url", None)
-    last = await db.email_logs.find_one(
-        {"channel": "discord"},
-        {"_id": 0, "status": 1, "error": 1, "event_key": 1, "created_at": 1},
-        sort=[("created_at", -1)],
-    )
-    if last:
-        s["last_status"] = last.get("status")
-        s["last_error"] = last.get("error")
-        s["last_event_key"] = last.get("event_key")
-        s["last_checked_at"] = last.get("created_at")
-    return s
-
-
-@settings_router.put("/discord")
-async def update_discord(body: DiscordSettings, me: dict = Depends(require_club_admin())):
-    db = get_db()
-    updates = {k: v for k, v in body.model_dump(exclude_unset=True).items() if v is not None}
-    clear_webhook = bool(updates.pop("clear_webhook", False))
-    unset = {"webhook_url": ""} if clear_webhook else {}
-    if "webhook_url" in updates:
-        updates["webhook_url"] = updates["webhook_url"].strip()
-    if "webhook_url" in updates and not updates["webhook_url"]:
-        updates.pop("webhook_url")
-    if "webhook_url" in updates:
-        from discord_service import is_valid_discord_webhook_url
-        if not is_valid_discord_webhook_url(updates["webhook_url"]):
-            raise HTTPException(400, "Ungültige Discord Webhook URL. Erlaubt sind https://discord.com/api/webhooks/... URLs.")
-        updates["webhook_url"] = encrypt_secret(updates["webhook_url"])
-    for key in ("username", "avatar_url"):
-        if key in updates and isinstance(updates[key], str):
-            updates[key] = updates[key].strip()
-    current = await db.settings.find_one({"id": "discord"}, {"_id": 0}) or {}
-    changed_fields = _changed_setting_fields(current, updates, unset)
-    if not changed_fields:
-        return {"ok": True, "changed": False}
-    updates["updated_at"] = now_utc().isoformat()
-    op = {"$set": updates, "$setOnInsert": {"id": "discord"}}
-    if unset:
-        op["$unset"] = unset
-    await db.settings.update_one(
-        {"id": "discord"}, op, upsert=True,
-    )
-    await _audit_settings_change(db, "settings.discord.update", "discord", me["id"], changed_fields)
-    return {"ok": True, "changed": True}
-
-
-@settings_router.post("/discord/test")
-async def discord_test(me: dict = Depends(require_club_admin())):
-    from discord_service import send_discord
-    res = await send_discord(
-        "THE LION SQUAD · Testnachricht",
-        "Diese Nachricht bestätigt, dass dein Discord-Webhook korrekt funktioniert. 🦁",
-        event_key="test",
-    )
-    return res
-
-
-# ---------- Seasons / Circuits ----------
-season_router = APIRouter(prefix="/api/seasons", tags=["seasons"])
-
-
-class SeasonCreate(BaseModel):
-    name: str
-    slug: Optional[str] = None
-    description: Optional[str] = None
-    kind: Literal["season", "circuit"] = "season"
-    tournament_ids: List[str] = []
-    f1_challenge_ids: List[str] = []
-    points_per_position: List[int] = [25, 18, 15, 12, 10, 8, 6, 4, 2, 1]
-    drop_worst: int = 0  # Streichresultate
-    bonus_points: dict = {}
-    start_date: Optional[datetime] = None
-    end_date: Optional[datetime] = None
-    banner_url: Optional[str] = None
-
-
-class SeasonUpdate(BaseModel):
-    name: Optional[str] = None
-    slug: Optional[str] = None
-    description: Optional[str] = None
-    kind: Optional[Literal["season", "circuit"]] = None
-    status: Optional[Literal["draft", "active", "completed", "archived"]] = None
-    tournament_ids: Optional[List[str]] = None
-    f1_challenge_ids: Optional[List[str]] = None
-    points_per_position: Optional[List[int]] = None
-    drop_worst: Optional[int] = None
-    start_date: Optional[datetime] = None
-    end_date: Optional[datetime] = None
-    banner_url: Optional[str] = None
-
-
-@season_router.get("")
-async def list_seasons():
-    db = get_db()
-    return await db.seasons.find({}, {"_id": 0}).sort("start_date", -1).to_list(200)
-
-
-@season_router.get("/active/featured")
-async def featured_season():
-    """Returns the most relevant active season + top 5 standings for public widgets."""
-    db = get_db()
-    s = await db.seasons.find_one({"status": "active"}, {"_id": 0},
-                                    sort=[("start_date", -1)])
-    if not s:
-        s = await db.seasons.find_one({}, {"_id": 0}, sort=[("start_date", -1)])
-    if not s:
-        return {"season": None, "standings": []}
-    # Reuse standings logic (season_standings defined below in this module)
-    lb = await season_standings(s.get("slug") or s["id"])
-    return {"season": lb["season"], "standings": (lb.get("standings") or [])[:5]}
-
-
-@season_router.get("/{slug_or_id}")
-async def get_season(slug_or_id: str):
-    db = get_db()
-    s, was_old_slug = await find_by_slug_or_history(db.seasons, slug_or_id, {"_id": 0})
-    if not s:
-        raise HTTPException(status_code=404, detail="Saison nicht gefunden")
-    if was_old_slug and s.get("slug"):
-        return RedirectResponse(url=f"/api/seasons/{s['slug']}", status_code=301)
-    tids, fids = await _resolve_season_sources(s)
-    s["tournaments"] = await db.tournaments.find({"id": {"$in": tids}}, {"_id": 0}).to_list(200)
-    s["f1_challenges"] = await db.f1_challenges.find({"id": {"$in": fids}},
-                                                       {"_id": 0}).to_list(200)
-    return s
-
-
-@season_router.post("")
-async def create_season(body: SeasonCreate, me: dict = Depends(require_admin())):
-    db = get_db()
-    doc = body.model_dump()
-    doc["slug"] = await unique_slug(db.seasons, doc.get("slug") or doc.get("name"), fallback="season")
-    for k in ["start_date", "end_date"]:
-        if doc.get(k):
-            doc[k] = doc[k].isoformat()
-    doc["id"] = new_id()
-    doc["status"] = "draft"
-    doc["created_at"] = now_utc().isoformat()
-    doc["updated_at"] = now_utc().isoformat()
-    doc["created_by"] = me["id"]
-    await db.seasons.insert_one(doc)
-    doc.pop("_id", None)
-    return doc
-
-
-@season_router.put("/{sid}")
-@season_router.patch("/{sid}")
-async def update_season(sid: str, body: SeasonUpdate, me: dict = Depends(require_admin())):
-    db = get_db()
-    current = await db.seasons.find_one({"$or": [{"id": sid}, {"slug": sid}]}, {"_id": 0})
-    if not current:
-        raise HTTPException(404, "Saison nicht gefunden")
-    nullable_fields = {"description", "banner_url", "start_date", "end_date"}
-    raw = body.model_dump(exclude_unset=True)
-    updates = {k: v for k, v in raw.items() if v is not None or k in nullable_fields}
-    slug_source = slug_source_for_update(raw, current, "name", fallback="season")
-    if slug_source is not None:
-        updates["slug"] = await unique_slug(db.seasons, slug_source, current_id=current["id"], fallback="season")
-        apply_slug_history(current, updates)
-    for k in ["start_date", "end_date"]:
-        if k in updates:
-            updates[k] = updates[k].isoformat() if updates[k] else None
-    updates["updated_at"] = now_utc().isoformat()
-    await db.seasons.update_one({"id": current["id"]}, {"$set": updates})
-    return await db.seasons.find_one({"id": current["id"]}, {"_id": 0})
-
-
-@season_router.delete("/{sid}")
-async def delete_season(sid: str, me: dict = Depends(require_admin())):
-    db = get_db()
-    await db.seasons.delete_one({"$or": [{"id": sid}, {"slug": sid}]})
-    return {"ok": True}
-
-
-# ---------- Phase 7: Jahreswertung v2 (Vereinsplattform spec) ----------
-@season_router.get("/v2/leaderboard")
-async def leaderboard_v2(
-    season_id: str | None = None,
-    only_members: bool = False,
-    only_community: bool = False,
-    rookie_only: bool = False,
-    teams: bool = False,
-    source_type: str | None = None,
-    limit: int = 100,
-):
-    """Aggregated standings using the Phase 7 points formula
-    (base × weight × participant_factor + bonus, with farming protection)."""
-    from services.season_service import aggregate_leaderboard
-    rows = await aggregate_leaderboard(
-        season_id=season_id,
-        only_members=only_members,
-        only_community=only_community,
-        rookie_only=rookie_only,
-        teams=teams,
-        source_type=source_type,
-        limit=limit,
-    )
-    for i, r in enumerate(rows):
-        r["rank"] = i + 1
-    return {"standings": rows}
-
-
-@season_router.get("/v2/me")
-async def my_season_points(me: dict = Depends(get_current_user)):
-    db = get_db()
-    season = await db.seasons.find_one({"status": "active"}, {"_id": 0})
-    if not season:
-        return {"season": None, "total": 0, "entries": []}
-    entries = await db.season_points.find(
-        {"season_id": season["id"], "user_id": me["id"]}, {"_id": 0},
-    ).sort("created_at", -1).to_list(500)
-    total = round(sum(e.get("total_points", 0) for e in entries), 1)
-    from services.season_service import _achievement_summaries, _summarise_point_entries
-    achievements = (await _achievement_summaries(db, [me["id"]])).get(me["id"], {})
-    return {
-        "season": season,
-        "total": total,
-        "season_points": total,
-        "entries": entries,
-        "source_breakdown": _summarise_point_entries(entries, max(int(season.get("drop_worst") or 0), 0)).get("source_breakdown", []),
-        "achievement_count": achievements.get("achievement_count", 0),
-        "achievement_points": achievements.get("achievement_points", 0),
-        "profile_points": achievements.get("achievement_points", 0),
-    }
-
-
-@season_router.post("/v2/award")
-async def award_points_admin(body: dict, me: dict = Depends(require_admin())):
-    """Admin: manually award season points to a user/team."""
-    from services.season_service import award_points
-    res = await award_points(
-        user_id=body.get("user_id"),
-        team_id=body.get("team_id"),
-        source_type=body.get("source_type", "custom"),
-        source_id=body.get("source_id"),
-        source_name=body.get("source_name", "Manuelle Vergabe"),
-        rank=body.get("rank"),
-        num_participants=int(body.get("num_participants", 1)),
-        weight=float(body["weight"]) if body.get("weight") is not None else None,
-        bonus=int(body.get("bonus", 0)),
-        bonus_reason=body.get("bonus_reason"),
-        farming_exempt=bool(body.get("farming_exempt", False)),
-    )
-    if res is None:
-        raise HTTPException(400, "Keine aktive Saison.")
-    return res
-
-
-@season_router.delete("/v2/entry/{entry_id}")
-async def delete_season_entry(entry_id: str, me: dict = Depends(require_admin())):
-    db = get_db()
-    res = await db.season_points.delete_one({"id": entry_id})
-    if res.deleted_count == 0:
-        raise HTTPException(404, "Eintrag nicht gefunden.")
-    return {"ok": True}
-
-
-async def _resolve_season_sources(s: dict) -> tuple[list[str], list[str]]:
-    """Resolve which tournaments + f1 challenges feed into this season.
-
-    Strategy:
-      - If `tournament_ids`/`f1_challenge_ids` are explicitly listed → use those.
-      - Otherwise auto-include every tournament/f1 challenge whose status is in
-        a relevant set AND whose start/created date falls inside the season
-        date range (start_date / end_date). Falls back to all if season has no
-        date range yet.
-    """
-    db = get_db()
-    explicit_t = list(s.get("tournament_ids") or [])
-    explicit_f = list(s.get("f1_challenge_ids") or [])
-    if explicit_t and explicit_f:
-        return explicit_t, explicit_f
-
-    # Build date filter (lenient: matches scheduled_at OR created_at fallback)
-    start = s.get("start_date")
-    end = s.get("end_date")
-    relevant_status = {"live", "completed", "results_published", "check_in", "scheduled"}
-
-    auto_t: list[str] = []
-    if not explicit_t:
-        async for t in db.tournaments.find({}, {"id": 1, "status": 1, "start_date": 1, "created_at": 1, "_id": 0}):
-            if t.get("status") not in relevant_status:
-                continue
-            ts = t.get("start_date") or t.get("created_at")
-            if start and end and ts and not (start <= ts <= end):
-                continue
-            auto_t.append(t["id"])
-    auto_f: list[str] = []
-    if not explicit_f:
-        async for f in db.f1_challenges.find({}, {"id": 1, "status": 1, "start_date": 1, "created_at": 1, "_id": 0}):
-            if f.get("status") not in relevant_status:
-                continue
-            ts = f.get("start_date") or f.get("created_at")
-            if start and end and ts and not (start <= ts <= end):
-                continue
-            auto_f.append(f["id"])
-
-    return (explicit_t or auto_t), (explicit_f or auto_f)
-
-
-@season_router.get("/{slug_or_id}/standings")
-async def season_standings(slug_or_id: str):
-    """Aggregate standings over all season point sources."""
-    db = get_db()
-    s, was_old_slug = await find_by_slug_or_history(db.seasons, slug_or_id, {"_id": 0})
-    if not s:
-        raise HTTPException(status_code=404, detail="Saison nicht gefunden")
-    if was_old_slug and s.get("slug"):
-        return RedirectResponse(url=f"/api/seasons/{s['slug']}/standings", status_code=301)
-    if await db.season_points.count_documents({"season_id": s["id"]}):
-        from services.season_service import aggregate_leaderboard
-        rows = await aggregate_leaderboard(season_id=s["id"], limit=500)
-        standings = []
-        for index, row in enumerate(rows):
-            standings.append({
-                **row,
-                "user_id": row.get("id"),
-                "display_name": row.get("display_name") or row.get("username") or "—",
-                "points": row.get("total_points", 0),
-                "events_count": row.get("events", 0),
-                "wins": row.get("wins", 0),
-                "rank": index + 1,
-            })
-        return {"season": s, "standings": standings}
-    points_system = s.get("points_per_position", [25, 18, 15, 12, 10, 8, 6, 4, 2, 1])
-    per_user_points: dict = {}  # user_id -> {points, events_points_list, wins}
-
-    def add_points(user_id, pts, won=False):
-        per_user_points.setdefault(user_id, {"user_id": user_id, "points": 0, "events": [], "wins": 0})
-        per_user_points[user_id]["events"].append(pts)
-        if won:
-            per_user_points[user_id]["wins"] += 1
-
-    tournament_ids, f1_ids = await _resolve_season_sources(s)
-
-    # Tournaments: use the same canonical placement/standings projections.
-    for tid in tournament_ids:
-        regs = await db.tournament_registrations.find(
-            {"tournament_id": tid, "status": {"$in": ["approved", "checked_in"]}},
-            {"_id": 0},
-        ).to_list(500)
-        reg_user_map = {r["id"]: r.get("user_id") for r in regs}
-        read_model = await load_competition_read_model(db, tid)
-        snapshot = read_model.structure_snapshot()
-        observe_structure_read(snapshot, surface="season_fallback")
-        tournament = await db.tournaments.find_one({"id": tid}, {"_id": 0, "format": 1}) or {}
-        groups = []
-        if tournament.get("format") == "groups":
-            groups = await db.tournament_groups.find(
-                {"tournament_id": tid}, {"_id": 0}
-            ).sort("order_index", 1).to_list(100)
-        standings = standings_for_structure(tournament, snapshot, regs, groups=groups)
-        rows = [
-            group_row
-            for item in standings
-            for group_row in (item.get("standings") or [])
-        ] if tournament.get("format") == "groups" else standings
-        for row in rows:
-            if "played" in row and not row.get("played"):
-                continue
-            uid = reg_user_map.get(row.get("registration_id"))
-            if not uid:
-                continue
-            pos = int(row.get("rank") or 999) - 1
-            pts = points_system[pos] if 0 <= pos < len(points_system) else 0
-            add_points(uid, pts, pos == 0)
-
-    # F1 Challenges: aggregate per-track then championship-style
-    for cid in f1_ids:
-        tracks = await db.f1_tracks.find({"challenge_id": cid}, {"_id": 0}).to_list(100)
-        for tr in tracks:
-            times = await db.f1_lap_times.find(
-                {
-                    "challenge_id": cid,
-                    "track_id": tr["id"],
-                    "is_invalid": {"$ne": True},
-                    "$or": [{"score_scope": {"$exists": False}}, {"score_scope": {"$ne": "club_reference"}}],
-                },
-                {"_id": 0},
-            ).to_list(5000)
-            best_per_user: dict = {}
-            for t in times:
-                eff = t["time_ms"] + int(t.get("penalty_seconds", 0) * 1000)
-                if t["user_id"] not in best_per_user or eff < best_per_user[t["user_id"]]:
-                    best_per_user[t["user_id"]] = eff
-            sorted_u = sorted(best_per_user.items(), key=lambda x: x[1])
-            for pos, (uid, _) in enumerate(sorted_u):
-                pts = points_system[pos] if pos < len(points_system) else 0
-                add_points(uid, pts, pos == 0)
-
-    # Apply drop_worst
-    drop_worst = s.get("drop_worst", 0)
-    for uid, st in per_user_points.items():
-        evts = sorted(st["events"], reverse=True)
-        if drop_worst and len(evts) > drop_worst:
-            evts = evts[: len(evts) - drop_worst]
-        st["points"] = sum(evts)
-        st["events_count"] = len(st["events"])
-
-    # Enrich users
-    user_ids = list(per_user_points.keys())
-    users = {u["id"]: u for u in await db.users.find(
-        {"id": {"$in": user_ids}}, {"_id": 0, "password_hash": 0, "mfa_secret": 0, "mfa_pending_secret": 0, "mfa_recovery_code_hashes": 0}).to_list(500)}
-    arr = []
-    for uid, st in per_user_points.items():
-        u = users.get(uid, {})
-        arr.append({**st,
-                     "display_name": u.get("display_name") or u.get("username") or "—",
-                     "username": u.get("username"),
-                     "avatar_url": u.get("avatar_url")})
-    arr.sort(key=lambda s: (s["points"], s["wins"]), reverse=True)
-    for i, s_ in enumerate(arr):
-        s_["rank"] = i + 1
-        s_.pop("events", None)
-    return {"season": s, "standings": arr}
-
-
-# ---------- Widgets ----------
-widget_router = APIRouter(prefix="/api/widgets", tags=["widgets"])
-
-
-async def _public_f1_challenge_or_404(slug_or_id: str, access: str | None = None) -> dict:
-    db = get_db()
-    c, _ = await find_by_slug_or_history(db.f1_challenges, slug_or_id, {"_id": 0})
-    if not c:
-        raise HTTPException(status_code=404)
-    access_link = await validate_access_link(db, access, "fastlap", c["id"], None, "view")
-    if not access_link and (c.get("status") == "draft" or (c.get("visibility") or "public") != "public"):
-        raise HTTPException(status_code=404)
-    return c
-
-
-async def _public_tournament_or_404(slug_or_id: str) -> dict:
-    db = get_db()
-    t, _ = await find_by_slug_or_history(db.tournaments, slug_or_id, {"_id": 0})
-    if (
-        not t
-        or t.get("status") == "draft"
-        or t.get("is_public") is False
-        or not await user_can_see(None, t.get("visibility") or "public")
-    ):
-        raise HTTPException(status_code=404)
-    return t
-
-
-def _public_registration(reg: dict) -> dict:
-    return {
-        "id": reg.get("id"),
-        "tournament_id": reg.get("tournament_id"),
-        "status": reg.get("status"),
-        "display_name": reg.get("display_name") or reg.get("ingame_name"),
-        "ingame_name": reg.get("ingame_name"),
-        "team_id": reg.get("team_id"),
-        "seed": reg.get("seed"),
-    }
-
-
-def _public_widget_legacy_match(match: dict) -> dict:
-    return {
-        key: value
-        for key, value in match.items()
-        if key not in {"admin_note", "reports", "disputes"}
-    }
-
-
-def _public_widget_stage_match(match: dict) -> dict:
-    public_fields = {
-        "id", "tournament_id", "stage_id", "stage_number", "stage_type",
-        "match_type", "match_key", "section", "round", "round_name", "order",
-        "slots", "results", "advancement", "status", "is_preview",
-        "generation_mode", "scheduled_at", "duration_minutes", "station_id",
-        "station_label", "station_name", "map", "best_of",
-    }
-    return {key: value for key, value in match.items() if key in public_fields}
-
-
-def _public_challenge_summary(challenge: dict) -> dict:
-    return {
-        "id": challenge.get("id"),
-        "slug": challenge.get("slug"),
-        "title": challenge.get("title"),
-        "status": challenge.get("status"),
-    }
-
-
-@widget_router.get("/tournament/{slug_or_id}/bracket")
-async def widget_bracket(slug_or_id: str):
-    """Read-only bracket data for widget embed."""
-    db = get_db()
-    t = await _public_tournament_or_404(slug_or_id)
-    read_model = await load_competition_read_model(db, t["id"])
-    structure = read_model.structure_snapshot()
-    observe_structure_read(structure, surface="widget")
-    regs = await db.tournament_registrations.find(
-        {"tournament_id": t["id"]},
-        {"_id": 0},
-    ).to_list(500)
-    return {
-        "tournament": {"id": t["id"], "title": t["title"], "format": t["format"], "status": t["status"]},
-        "matches": [_public_widget_legacy_match(match) for match in read_model.legacy_matches],
-        "matches_v2": [_public_widget_stage_match(match) for match in read_model.stage_matches],
-        "stages": read_model.stages,
-        "engine": "stage" if read_model.stages or read_model.stage_matches else "legacy",
-        "structure": structure,
-        "registrations": [_public_registration(r) for r in regs],
-    }
-
-
-@widget_router.get("/f1/{slug_or_id}/leaderboard")
-async def widget_f1(slug_or_id: str, track_id: Optional[str] = None):
-    db = get_db()
-    c = await _public_f1_challenge_or_404(slug_or_id)
-    if not track_id:
-        first = await db.f1_tracks.find_one({"challenge_id": c["id"]}, {"_id": 0}, sort=[("order_index", 1)])
-        if not first:
-            return {"challenge": _public_challenge_summary(c), "track": None, "entries": []}
-        track_id = first["id"]
-    # reuse f1 leaderboard logic (inline-light)
-    track = await db.f1_tracks.find_one({"id": track_id}, {"_id": 0})
-    times = await db.f1_lap_times.find(
-        {"challenge_id": c["id"], "track_id": track_id, "is_invalid": {"$ne": True}},
-        {"_id": 0, "admin_note": 0, "proof_url": 0},
-    ).to_list(5000)
-    best_per_user = {}
-    for t in times:
-        eff = t["time_ms"] + int(t.get("penalty_seconds", 0) * 1000)
-        if t["user_id"] not in best_per_user or eff < best_per_user[t["user_id"]]["effective_ms"]:
-            best_per_user[t["user_id"]] = {**t, "effective_ms": eff}
-    user_ids = list(best_per_user.keys())
-    users = {u["id"]: u for u in await db.users.find(
-        {"id": {"$in": user_ids}}, {"_id": 0, "password_hash": 0, "email": 0, "mfa_secret": 0, "mfa_pending_secret": 0, "mfa_recovery_code_hashes": 0}).to_list(500)}
-    entries = []
-    for uid, tr in best_per_user.items():
-        u = users.get(uid, {})
-        m = tr["effective_ms"]
-        entries.append({"display_name": u.get("display_name") or u.get("username") or "—",
-                         "time_ms": m,
-                         "time_str": f"{m//60000}:{(m%60000)//1000:02d}.{m%1000:03d}"})
-    entries.sort(key=lambda e: e["time_ms"])
-    for i, e in enumerate(entries):
-        e["rank"] = i + 1
-        e["gap_ms"] = e["time_ms"] - entries[0]["time_ms"] if i > 0 else 0
-        e["gap_str"] = f"+{e['gap_ms']/1000:.3f}s" if i > 0 else ""
-    return {"challenge": _public_challenge_summary(c),
-            "track": track, "entries": entries}
-
-
-# ---------- DSGVO ----------
-dsgvo_router = APIRouter(prefix="/api/dsgvo", tags=["dsgvo"])
-
-
-async def _user_data_export(db, user_id: str) -> dict:
-    user = await db.users.find_one(
-        {"id": user_id},
-        {"_id": 0, "password_hash": 0, "mfa_secret": 0, "mfa_pending_secret": 0,
-         "mfa_recovery_code_hashes": 0},
-    )
-    if not user:
-        raise HTTPException(404, "Account nicht gefunden")
-    email = user.get("email", "")
-    async def rows(collection, query, limit=5000):
-        return await collection.find(query, {"_id": 0}).to_list(limit)
-
-    registrations = await rows(db.tournament_registrations, {"user_id": user_id})
-    return {
-        "format_version": 2,
-        "exported_at": now_utc().isoformat(),
-        "user": user,
-        "consent_records": await rows(db.consent_records, {"user_id": user_id}),
-        "membership": await db.memberships.find_one({"user_id": user_id}, {"_id": 0}),
-        "social_accounts": await rows(db.user_socials, {"user_id": user_id}),
-        "tournament_registrations": registrations,
-        "competition_matches": await registration_match_snapshot(
-            db, [row.get("id") for row in registrations if row.get("id")],
-        ),
-        "event_registrations": await rows(db.event_registrations, {"user_id": user_id}),
-        "f1_lap_times": await rows(db.f1_lap_times, {"user_id": user_id}),
-        "teams": await rows(db.teams, {"$or": [{"member_ids": user_id}, {"leader_id": user_id}]}),
-        "team_memberships": await rows(db.team_members, {"user_id": user_id}),
-        "team_invites": await rows(db.team_invites, {"user_id": user_id}),
-        "achievements": await rows(db.user_achievements, {"user_id": user_id}),
-        "season_points": await rows(db.season_points, {"user_id": user_id}),
-        "prize_pickups": await rows(db.prize_pickups, {"user_id": user_id}),
-        "notifications": await rows(db.notifications, {"user_id": user_id}),
-        "direct_messages": await rows(db.direct_messages, {"$or": [{"sender_id": user_id}, {"recipient_id": user_id}]}),
-        "friendships": await rows(db.friendships, {"$or": [{"requester_id": user_id}, {"recipient_id": user_id}]}),
-        "blocks": await rows(db.user_blocks, {"$or": [{"blocker_id": user_id}, {"blocked_id": user_id}]}),
-        "moderation_reports": await rows(db.user_reports, {"$or": [{"reporter_id": user_id}, {"target_user_id": user_id}]}),
-        "email_logs": await rows(db.email_logs, {"to": email}),
-        "mobile_devices": await rows(db.mobile_push_tokens, {"user_id": user_id}),
-        "mobile_client_logs": await rows(db.mobile_client_logs, {"user_id": user_id}),
-        "audit_trail": await rows(db.audit_logs, {"$or": [{"actor_id": user_id}, {"target_id": user_id}]}),
-    }
-
-
-async def _anonymize_user_data(db, user_id: str, actor_id: str, action: str) -> None:
-    user = await db.users.find_one({"id": user_id})
-    if not user:
-        raise HTTPException(404, "Account nicht gefunden")
-    now = now_utc().isoformat()
-    anonymous_email = f"deleted_{user_id[:12]}@deleted.invalid"
-    anonymous_username = f"deleted_{user_id[:12]}"
-    await db.users.update_one({"id": user_id}, {
-        "$set": {
-            "email": anonymous_email, "username": anonymous_username, "display_name": "Gelöschter User",
-            "first_name": None, "last_name": None, "nickname": None, "birth_date": None, "gender": None,
-            "bio": None, "discord_name": None, "discord_id": None, "switch_code": None, "steam_id": None,
-            "epic_id": None, "psn_id": None, "xbox_id": None, "riot_id": None, "game_ids": {},
-            "country": None, "state": None, "city": None, "avatar_url": None, "banner_url": None,
-            "twitch_handle": None, "youtube_handle": None, "instagram_handle": None,
-            "newsletter_consent": False, "notification_preferences": {}, "privacy_public_profile": False,
-            "google_linked": False, "password_login_available": False, "email_verified": False,
-            "is_active": False, "is_banned": True, "mfa_enabled": False,
-            "password_hash": "!disabled", "anonymized_at": now, "updated_at": now,
-        },
-        "$unset": {
-            "google_id": "", "google_email": "", "mfa_secret": "", "mfa_pending_secret": "",
-            "mfa_pending_created_at": "", "mfa_recovery_code_hashes": "",
-        },
-    })
-    for collection in (db.refresh_tokens, db.auth_sessions, db.email_verification_tokens,
-                       db.password_reset_tokens, db.mfa_login_challenges, db.mobile_push_tokens,
-                       db.mobile_client_logs, db.notifications, db.user_socials,
-                       db.passkeys, db.passkey_challenges):
-        await collection.delete_many({"user_id": user_id})
-    await db.friendships.delete_many({"$or": [{"requester_id": user_id}, {"recipient_id": user_id}]})
-    await db.user_blocks.delete_many({"$or": [{"blocker_id": user_id}, {"blocked_id": user_id}]})
-    await db.direct_messages.update_many({"sender_id": user_id}, {"$set": {"message": "[Nachricht gelöscht]", "sender_anonymized": True}})
-    await db.team_chat_messages.update_many({"user_id": user_id}, {"$set": {"message": "[Nachricht gelöscht]", "author_anonymized": True}})
-    await db.match_chat_messages.update_many({"user_id": user_id}, {"$set": {"message": "[Nachricht gelöscht]", "author_anonymized": True}})
-    await db.email_logs.update_many({"to": user.get("email")}, {"$set": {"to": anonymous_email, "recipient_anonymized": True}})
-    await db.memberships.update_many({"user_id": user_id}, {"$set": {
-        "email": anonymous_email, "first_name": None, "last_name": None, "phone": None,
-        "address": None, "member_status": "former", "updated_at": now,
-    }})
-    await db.audit_logs.insert_one({
-        "id": new_id(), "action": action, "actor_id": actor_id, "target_id": user_id,
-        "data": {"personal_data_removed": True}, "created_at": now,
-    })
-
-
-@dsgvo_router.get("/export-my-data")
-async def export_my_data(me: dict = Depends(get_current_user)):
-    db = get_db()
-    return await _user_data_export(db, me["id"])
-
-
-@dsgvo_router.post("/anonymize-me")
-async def anonymize_me(me: dict = Depends(get_current_user)):
-    """Anonymize own account but keep tournament history for statistical integrity."""
-    db = get_db()
-    await _anonymize_user_data(db, me["id"], me["id"], "user.self_anonymize")
-    return {"ok": True}
-
-
-@dsgvo_router.post("/admin/anonymize/{user_id}")
-async def admin_anonymize(user_id: str, me: dict = Depends(require_super())):
-    db = get_db()
-    target = await db.users.find_one({"id": user_id}, {"_id": 0, "role": 1})
-    if target and target.get("role") == "superadmin" and user_id != me["id"]:
-        raise HTTPException(403, "Andere Superadmins dürfen nicht anonymisiert werden.")
-    await _anonymize_user_data(db, user_id, me["id"], "user.admin_anonymize")
-    return {"ok": True}
-
-
-# ---------- PDF Exports ----------
-pdf_router = APIRouter(prefix="/api/exports", tags=["exports"])
-
-
-def _pdf_response(data: bytes, filename: str):
-    safe_filename = _pdf_filename(filename)
-    return StreamingResponse(
-        io.BytesIO(data),
-        media_type="application/pdf",
-        headers={"Content-Disposition": f'attachment; filename="{safe_filename}"'},
-    )
-
-
-def _pdf_filename_part(*values: str | None, fallback: str = "export") -> str:
-    for value in values:
-        raw = str(value or "").strip()
-        if not raw:
-            continue
-        safe = re.sub(r"[^A-Za-z0-9._-]+", "_", raw).strip("._-")
-        if safe:
-            return safe[:90]
-    return fallback
-
-
-def _pdf_filename(filename: str) -> str:
-    raw = str(filename or "").strip()
-    stem = raw[:-4] if raw.lower().endswith(".pdf") else raw
-    return _pdf_filename_part(stem, fallback="export") + ".pdf"
-
-
-_PDF_SPONSOR_TIER_ORDER = {"main": 0, "platinum": 1, "gold": 2, "silver": 3, "bronze": 4}
-
-
-def _sponsor_date(value) -> str | None:
-    raw = str(value or "").strip()
-    return raw[:10] if raw else None
-
-
-def _sponsor_active_for_pdf(sponsor: dict) -> bool:
-    if sponsor.get("is_active") is False or sponsor.get("show_on_pdf") is not True:
-        return False
-    status = str(sponsor.get("contract_status") or "active").strip().lower()
-    if status in {"paused", "cancelled"}:
-        return False
-    today = datetime.now(timezone.utc).date().isoformat()
-    start = _sponsor_date(sponsor.get("contract_start"))
-    end = _sponsor_date(sponsor.get("contract_end"))
-    return not ((start and start > today) or (end and end < today))
-
-
-async def _pdf_sponsors(db):
-    rows = await db.sponsors.find(
-        {"is_active": {"$ne": False}, "show_on_pdf": True},
-        {"_id": 0, "name": 1, "logo_url": 1, "link": 1, "tier": 1, "order_index": 1,
-         "is_active": 1, "contract_status": 1, "contract_start": 1, "contract_end": 1, "show_on_pdf": 1},
-    ).to_list(50)
-    rows = [s for s in rows if _sponsor_active_for_pdf(s)]
-    rows.sort(key=lambda s: (_PDF_SPONSOR_TIER_ORDER.get(s.get("tier"), 99), s.get("order_index") or 0, s.get("name") or ""))
-    return rows
-
-
-async def _pdf_branding(db):
-    branding = await db.settings.find_one(
-        {"id": "branding"},
-        {
-            "_id": 0,
-            "logo_url": 1,
-            "logo_light_url": 1,
-            "logo_dark_url": 1,
-            "mascot_url": 1,
-            "qr_logo_url": 1,
-            "favicon_url": 1,
-            "favicon_light_url": 1,
-            "favicon_dark_url": 1,
-            "domain": 1,
-            "site_title": 1,
-            "club_name": 1,
-        },
-    ) or {}
-    if not branding.get("logo_url"):
-        branding["logo_url"] = branding.get("logo_dark_url") or branding.get("logo_light_url")
-    return branding
-
-
-def _pdf_public_base(branding: dict) -> str:
-    base = str((branding or {}).get("domain") or "https://lionsquad.at").strip().rstrip("/")
-    if base and not base.startswith(("http://", "https://")):
-        base = "https://" + base
-    return base or "https://lionsquad.at"
-
-
-def _pdf_absolute_url(path_or_url: str, branding: dict) -> str:
-    raw = str(path_or_url or "").strip()
-    if raw.startswith(("http://", "https://")):
-        return raw
-    base = _pdf_public_base(branding)
-    if not raw.startswith("/"):
-        raw = "/" + raw
-    return base + raw
-
-
-def _is_export_staff(user: dict | None) -> bool:
-    return bool(user and user.get("role") in STAFF_EXPORT_ROLES)
-
-
-async def _result_export_allowed(item: dict, user: dict | None) -> bool:
-    if _is_export_staff(user):
-        return True
-    if item.get("status") not in RESULT_EXPORT_STATUSES:
-        return False
-    if item.get("is_public") is False:
-        return False
-    return await user_can_see(user, item.get("visibility") or "public")
-
-
-async def _result_export_allowed_for_request(
-    db,
-    item: dict,
-    user: dict | None,
-    access: str | None,
-    resource_type: str,
-) -> bool:
-    if await _result_export_allowed(item, user):
-        return True
-    if not access or item.get("status") not in RESULT_EXPORT_STATUSES:
-        return False
-    return bool(await validate_access_link(db, access, resource_type, item.get("id"), user, "view"))
-
-
-def _certificate_source_from_item(item: dict, subtitle: str) -> dict:
-    return {
-        "title": item.get("title") or item.get("name") or "THE LION SQUAD",
-        "subtitle": subtitle,
-        "certificate_image_url": item.get("certificate_image_url"),
-        "banner_url": item.get("banner_url") or item.get("share_banner_url"),
-        "image_url": item.get("image_url"),
-        "seo_image_url": item.get("seo_image_url") or item.get("seo_banner_url"),
-    }
-
-
-def _top_certificate_rows(rows: list[dict], limit: int = 4) -> list[dict]:
-    result = []
-    for row in rows or []:
-        try:
-            rank = int(row.get("rank") or 0)
-        except (TypeError, ValueError):
-            rank = 0
-        if 1 <= rank <= limit:
-            result.append(row)
-    return sorted(result, key=lambda row: int(row.get("rank") or 999))
-
-
-def _tournament_certificate_metrics(row: dict) -> list[dict]:
-    metrics = [
-        {"label": "Siege", "value": row.get("won", row.get("wins", 0))},
-        {"label": "Niederlagen", "value": row.get("lost", row.get("losses", 0))},
-    ]
-    if row.get("points") is not None:
-        metrics.append({"label": "Punkte", "value": row.get("points")})
-    elif row.get("furthest_round") is not None:
-        metrics.append({"label": "Runde", "value": row.get("furthest_round")})
-    return metrics
-
-
-def _f1_track_certificate_metrics(row: dict, track: dict | None) -> list[dict]:
-    return [
-        {"label": "Strecke", "value": (track or {}).get("name") or "Fast Lap"},
-        {"label": "Beste Zeit", "value": row.get("time_str") or "—"},
-        {"label": "Abstand", "value": row.get("gap_str") or ("Führender" if row.get("rank") == 1 else "—")},
-        {"label": "Versuche", "value": row.get("attempts", 0)},
-    ]
-
-
-def _f1_championship_certificate_metrics(row: dict) -> list[dict]:
-    return [
-        {"label": "Punkte", "value": row.get("points", 0)},
-        {"label": "Siege", "value": row.get("wins", 0)},
-        {"label": "Rennen", "value": row.get("races", 0)},
-    ]
-
-
-@pdf_router.get("/qr/sign.pdf")
-async def pdf_qr_sign_export(
-    url: str,
-    title: str = "THE LION SQUAD",
-    subtitle: str = "",
-    eyebrow: str = "QR CODE",
-    me: dict = Depends(require_role("moderator")),
-):
-    db = get_db()
-    sponsors = await _pdf_sponsors(db)
-    branding = await _pdf_branding(db)
-    target_url = _pdf_absolute_url(url, branding)
-    return _pdf_response(
-        pdf_qr_sign(title, target_url, subtitle=subtitle, eyebrow=eyebrow, pdf_sponsors=sponsors, pdf_branding=branding),
-        "qr_schild.pdf",
-    )
-
-
-@pdf_router.get("/tournaments/{slug_or_id}/participants.pdf")
-async def pdf_tournament_participants(slug_or_id: str, me: dict = Depends(require_role("moderator"))):
-    db = get_db()
-    t = await db.tournaments.find_one({"$or": [{"id": slug_or_id}, {"slug": slug_or_id}]}, {"_id": 0})
-    if not t:
-        raise HTTPException(status_code=404)
-    regs = await db.tournament_registrations.find({"tournament_id": t["id"]}, {"_id": 0}).to_list(500)
-    team_ids = list({r["team_id"] for r in regs if r.get("team_id")})
-    teams = {x["id"]: x for x in await db.teams.find({"id": {"$in": team_ids}}, {"_id": 0}).to_list(500)}
-    for r in regs:
-        if r.get("team_id"):
-            r["team"] = teams.get(r["team_id"], {})
-    sponsors = await _pdf_sponsors(db)
-    branding = await _pdf_branding(db)
-    data = pdf_participants(t, regs, pdf_sponsors=sponsors, pdf_branding=branding)
-    file_id = _pdf_filename_part(t.get("slug"), t.get("id"), slug_or_id, fallback="turnier")
-    return _pdf_response(data, f"teilnehmer_{file_id}.pdf")
-
-
-@pdf_router.get("/tournaments/{slug_or_id}/checkin.pdf")
-async def pdf_tournament_checkin(slug_or_id: str, me: dict = Depends(require_role("moderator"))):
-    db = get_db()
-    t = await db.tournaments.find_one({"$or": [{"id": slug_or_id}, {"slug": slug_or_id}]}, {"_id": 0})
-    if not t:
-        raise HTTPException(status_code=404)
-    regs = await db.tournament_registrations.find({"tournament_id": t["id"]}, {"_id": 0}).to_list(500)
-    sponsors = await _pdf_sponsors(db)
-    branding = await _pdf_branding(db)
-    file_id = _pdf_filename_part(t.get("slug"), t.get("id"), slug_or_id, fallback="turnier")
-    return _pdf_response(pdf_checkin(t, regs, pdf_sponsors=sponsors, pdf_branding=branding), f"checkin_{file_id}.pdf")
-
-
-@pdf_router.get("/tournaments/{slug_or_id}/registration-qr.pdf")
-async def pdf_tournament_registration_qr(slug_or_id: str, me: dict = Depends(require_role("moderator"))):
-    db = get_db()
-    t = await db.tournaments.find_one({"$or": [{"id": slug_or_id}, {"slug": slug_or_id}]}, {"_id": 0})
-    if not t:
-        raise HTTPException(status_code=404)
-    sponsors = await _pdf_sponsors(db)
-    branding = await _pdf_branding(db)
-    public_id = _pdf_filename_part(t.get("slug"), t.get("id"), slug_or_id, fallback="turnier")
-    url = _pdf_absolute_url(f"/tournaments/{public_id}", branding)
-    subtitle = "Anmeldung, Check-in und Turnierinfos"
-    return _pdf_response(
-        pdf_qr_sign(t.get("title") or "Turnier", url, subtitle=subtitle, eyebrow="Turnier-Anmeldung", pdf_sponsors=sponsors, pdf_branding=branding),
-        f"anmeldung_qr_{public_id}.pdf",
-    )
-
-
-@pdf_router.get("/tournaments/{slug_or_id}/matches.pdf")
-async def pdf_tournament_matches(slug_or_id: str, me: dict = Depends(require_role("moderator"))):
-    db = get_db()
-    t = await db.tournaments.find_one({"$or": [{"id": slug_or_id}, {"slug": slug_or_id}]}, {"_id": 0})
-    if not t:
-        raise HTTPException(status_code=404)
-    read_model = await load_competition_read_model(db, t["id"])
-    structure = read_model.structure_snapshot()
-    observe_structure_read(structure, surface="match_pdf")
-    matches = structure["matches"]
-    regs = await db.tournament_registrations.find({"tournament_id": t["id"]}, {"_id": 0}).to_list(500)
-    reg_map = {r["id"]: r for r in regs}
-    sponsors = await _pdf_sponsors(db)
-    branding = await _pdf_branding(db)
-    file_id = _pdf_filename_part(t.get("slug"), t.get("id"), slug_or_id, fallback="turnier")
-    return _pdf_response(pdf_matches(t, matches, reg_map, pdf_sponsors=sponsors, pdf_branding=branding), f"matches_{file_id}.pdf")
-
-
-@pdf_router.get("/tournaments/{slug_or_id}/stations.pdf")
-async def pdf_tournament_station_signs(
-    slug_or_id: str,
-    orientation: Literal["portrait", "landscape"] = "portrait",
-    me: dict = Depends(require_role("moderator")),
-):
-    db = get_db()
-    t = await db.tournaments.find_one({"$or": [{"id": slug_or_id}, {"slug": slug_or_id}]}, {"_id": 0})
-    if not t:
-        raise HTTPException(status_code=404)
-    stations = await db.stations.find({"tournament_id": t["id"]}, {"_id": 0}).sort("name", 1).to_list(500)
-    sponsors = await _pdf_sponsors(db)
-    branding = await _pdf_branding(db)
-    return _pdf_response(
-        pdf_station_signs(t, stations, pdf_sponsors=sponsors, pdf_branding=branding, orientation=orientation),
-        f"stationen_{orientation}_{_pdf_filename_part(t.get('slug'), t.get('id'), slug_or_id, fallback='turnier')}.pdf",
-    )
-
-
-@pdf_router.get("/tournaments/{slug_or_id}/standings.pdf")
-async def pdf_tournament_standings(
-    slug_or_id: str,
-    access: Optional[str] = None,
-    user: dict | None = Depends(get_optional_user),
-):
-    db = get_db()
-    t = await db.tournaments.find_one({"$or": [{"id": slug_or_id}, {"slug": slug_or_id}]}, {"_id": 0})
-    if not t:
-        raise HTTPException(status_code=404)
-    if not await _result_export_allowed_for_request(db, t, user, access, "tournament"):
-        raise HTTPException(status_code=403, detail="Ergebnis-PDF ist erst nach Turnierende öffentlich.")
-    # Reuse standings logic
-    from routes.tournament_routes import standings as st_fn
-    rows = await st_fn(t["id"], access=access, user=user)
-    sponsors = await _pdf_sponsors(db)
-    branding = await _pdf_branding(db)
-    file_id = _pdf_filename_part(t.get("slug"), t.get("id"), slug_or_id, fallback="turnier")
-    return _pdf_response(pdf_standings(t, rows, pdf_sponsors=sponsors, pdf_branding=branding), f"standings_{file_id}.pdf")
-
-
-@pdf_router.get("/tournaments/{slug_or_id}/certificates.pdf")
-async def pdf_tournament_certificates(
-    slug_or_id: str,
-    access: Optional[str] = None,
-    user: dict | None = Depends(get_optional_user),
-):
-    db = get_db()
-    t = await db.tournaments.find_one({"$or": [{"id": slug_or_id}, {"slug": slug_or_id}]}, {"_id": 0})
-    if not t:
-        raise HTTPException(status_code=404)
-    if not await _result_export_allowed_for_request(db, t, user, access, "tournament"):
-        raise HTTPException(status_code=403, detail="Urkunden sind erst nach Turnierende öffentlich.")
-    from routes.tournament_routes import standings as st_fn
-    rows = _top_certificate_rows(await st_fn(t["id"], access=access, user=user))
-    if not rows:
-        raise HTTPException(status_code=404, detail="Keine Top-4-Platzierungen für Urkunden gefunden.")
-    sponsors = await _pdf_sponsors(db)
-    branding = await _pdf_branding(db)
-    source = _certificate_source_from_item(t, "Turnier")
-    certificates = [
-        {
-            "source": source,
-            "row": row,
-            "category": "Gesamtwertung",
-            "metrics": _tournament_certificate_metrics(row),
-        }
-        for row in rows
-    ]
-    file_id = _pdf_filename_part(t.get("slug"), t.get("id"), slug_or_id, fallback="turnier")
-    return _pdf_response(
-        pdf_certificates(certificates, pdf_sponsors=sponsors, pdf_branding=branding),
-        f"urkunden_{file_id}.pdf",
-    )
-
-
-@pdf_router.get("/tournaments/{slug_or_id}/certificates/{registration_id}.pdf")
-async def pdf_tournament_certificate(
-    slug_or_id: str,
-    registration_id: str,
-    access: Optional[str] = None,
-    user: dict | None = Depends(get_optional_user),
-):
-    db = get_db()
-    t = await db.tournaments.find_one({"$or": [{"id": slug_or_id}, {"slug": slug_or_id}]}, {"_id": 0})
-    if not t:
-        raise HTTPException(status_code=404)
-    if not await _result_export_allowed_for_request(db, t, user, access, "tournament"):
-        raise HTTPException(status_code=403, detail="Urkunden sind erst nach Turnierende öffentlich.")
-    from routes.tournament_routes import standings as st_fn
-    rows = await st_fn(t["id"], access=access, user=user)
-    row = next((item for item in rows if item.get("registration_id") == registration_id), None)
-    if not row:
-        raise HTTPException(status_code=404, detail="Platzierung nicht gefunden.")
-    if row not in _top_certificate_rows(rows):
-        raise HTTPException(status_code=404, detail="Urkunden werden für die ersten vier Plätze erzeugt.")
-    sponsors = await _pdf_sponsors(db)
-    branding = await _pdf_branding(db)
-    file_id = _pdf_filename_part(t.get("slug"), t.get("id"), slug_or_id, fallback="turnier")
-    player_id = _pdf_filename_part(row.get("display_name"), registration_id, fallback="spieler")
-    return _pdf_response(
-        pdf_certificate(
-            _certificate_source_from_item(t, "Turnier"),
-            row,
-            category="Gesamtwertung",
-            metrics=_tournament_certificate_metrics(row),
-            pdf_sponsors=sponsors,
-            pdf_branding=branding,
-        ),
-        f"urkunde_{file_id}_{player_id}.pdf",
-    )
-
-
-@pdf_router.get("/f1/{slug_or_id}/leaderboard.pdf")
-async def pdf_f1_lb(slug_or_id: str, track_id: Optional[str] = None, access: Optional[str] = None, user: dict | None = Depends(get_optional_user)):
-    db = get_db()
-    from routes.f1_routes import _get_visible_challenge, leaderboard as f1_lb
-    c = await _get_visible_challenge(slug_or_id, user, access=access)
-    if not await _result_export_allowed(c, user):
-        raise HTTPException(status_code=403, detail="Ergebnis-PDF ist erst nach Challenge-Ende öffentlich.")
-    lb = await f1_lb(c["id"], track_id, access, user)
-    sponsors = await _pdf_sponsors(db)
-    branding = await _pdf_branding(db)
-    file_id = _pdf_filename_part(c.get("slug"), c.get("id"), slug_or_id, fallback="f1")
-    return _pdf_response(pdf_f1_leaderboard(c, lb.get("track"), lb.get("entries", []), pdf_sponsors=sponsors, pdf_branding=branding),
-                          f"f1_{file_id}.pdf")
-
-
-@pdf_router.get("/f1/{slug_or_id}/certificates.pdf")
-async def pdf_f1_certificates(slug_or_id: str, track_id: Optional[str] = None, access: Optional[str] = None, user: dict | None = Depends(get_optional_user)):
-    db = get_db()
-    from routes.f1_routes import _get_visible_challenge, leaderboard as f1_lb
-    c = await _get_visible_challenge(slug_or_id, user, access=access)
-    if not await _result_export_allowed(c, user):
-        raise HTTPException(status_code=403, detail="Urkunden sind erst nach Challenge-Ende öffentlich.")
-    lb = await f1_lb(c["id"], track_id, access, user)
-    rows = _top_certificate_rows(lb.get("entries") or [])
-    if not rows:
-        raise HTTPException(status_code=404, detail="Keine Top-4-Platzierungen für Urkunden gefunden.")
-    sponsors = await _pdf_sponsors(db)
-    branding = await _pdf_branding(db)
-    track = lb.get("track") or {}
-    source = _certificate_source_from_item({**c, "image_url": track.get("image_url") or c.get("banner_url")}, "Fast Lap")
-    certificates = [
-        {
-            "source": source,
-            "row": row,
-            "category": f"Streckenwertung · {track.get('name') or 'Fast Lap'}",
-            "metrics": _f1_track_certificate_metrics(row, track),
-        }
-        for row in rows
-    ]
-    file_id = _pdf_filename_part(c.get("slug"), c.get("id"), slug_or_id, fallback="f1")
-    return _pdf_response(
-        pdf_certificates(certificates, pdf_sponsors=sponsors, pdf_branding=branding),
-        f"urkunden_f1_{file_id}.pdf",
-    )
-
-
-@pdf_router.get("/f1/{slug_or_id}/certificates/{user_id}.pdf")
-async def pdf_f1_certificate(slug_or_id: str, user_id: str, track_id: Optional[str] = None, access: Optional[str] = None, user: dict | None = Depends(get_optional_user)):
-    db = get_db()
-    from routes.f1_routes import _get_visible_challenge, leaderboard as f1_lb
-    c = await _get_visible_challenge(slug_or_id, user, access=access)
-    if not await _result_export_allowed(c, user):
-        raise HTTPException(status_code=403, detail="Urkunden sind erst nach Challenge-Ende öffentlich.")
-    lb = await f1_lb(c["id"], track_id, access, user)
-    rows = lb.get("entries") or []
-    row = next((item for item in rows if item.get("user_id") == user_id), None)
-    if not row:
-        raise HTTPException(status_code=404, detail="Platzierung nicht gefunden.")
-    if row not in _top_certificate_rows(rows):
-        raise HTTPException(status_code=404, detail="Urkunden werden für die ersten vier Plätze erzeugt.")
-    sponsors = await _pdf_sponsors(db)
-    branding = await _pdf_branding(db)
-    track = lb.get("track") or {}
-    file_id = _pdf_filename_part(c.get("slug"), c.get("id"), slug_or_id, fallback="f1")
-    player_id = _pdf_filename_part(row.get("display_name"), user_id, fallback="fahrer")
-    return _pdf_response(
-        pdf_certificate(
-            _certificate_source_from_item({**c, "image_url": track.get("image_url") or c.get("banner_url")}, "Fast Lap"),
-            row,
-            category=f"Streckenwertung · {track.get('name') or 'Fast Lap'}",
-            metrics=_f1_track_certificate_metrics(row, track),
-            pdf_sponsors=sponsors,
-            pdf_branding=branding,
-        ),
-        f"urkunde_f1_{file_id}_{player_id}.pdf",
-    )
-
-
-@pdf_router.get("/f1/{slug_or_id}/championship.pdf")
-async def pdf_f1_championship(slug_or_id: str, access: Optional[str] = None, user: dict | None = Depends(get_optional_user)):
-    db = get_db()
-    from routes.f1_routes import _get_visible_challenge, championship_standings as f1_champ
-    c = await _get_visible_challenge(slug_or_id, user, access=access)
-    if not await _result_export_allowed(c, user):
-        raise HTTPException(status_code=403, detail="Ergebnis-PDF ist erst nach Challenge-Ende öffentlich.")
-    cs = await f1_champ(c["id"], access, user)
-    # Reuse standings PDF shape
-    rows = [{"rank": r["rank"], "display_name": r["display_name"],
-             "won": r.get("wins", 0), "lost": (r.get("races", 0) - r.get("wins", 0)),
-             "points": r.get("points", 0)} for r in (cs.get("standings") or [])]
-    file_id = _pdf_filename_part(c.get("slug"), c.get("id"), slug_or_id, fallback="f1")
-    fake_tournament = {"title": (c.get("title") or "F1") + " · Championship", "slug": file_id}
-    sponsors = await _pdf_sponsors(db)
-    branding = await _pdf_branding(db)
-    return _pdf_response(pdf_standings(fake_tournament, rows, pdf_sponsors=sponsors, pdf_branding=branding),
-                          f"f1_championship_{file_id}.pdf")
-
-
-@pdf_router.get("/f1/{slug_or_id}/championship-certificates.pdf")
-async def pdf_f1_championship_certificates(slug_or_id: str, access: Optional[str] = None, user: dict | None = Depends(get_optional_user)):
-    db = get_db()
-    from routes.f1_routes import _get_visible_challenge, championship_standings as f1_champ
-    c = await _get_visible_challenge(slug_or_id, user, access=access)
-    if not await _result_export_allowed(c, user):
-        raise HTTPException(status_code=403, detail="Urkunden sind erst nach Challenge-Ende öffentlich.")
-    cs = await f1_champ(c["id"], access, user)
-    rows = _top_certificate_rows(cs.get("standings") or [])
-    if not rows:
-        raise HTTPException(status_code=404, detail="Keine Top-4-Platzierungen für Urkunden gefunden.")
-    sponsors = await _pdf_sponsors(db)
-    branding = await _pdf_branding(db)
-    source = _certificate_source_from_item(c, "Fast Lap Championship")
-    certificates = [
-        {
-            "source": source,
-            "row": row,
-            "category": "Gesamtwertung",
-            "metrics": _f1_championship_certificate_metrics(row),
-        }
-        for row in rows
-    ]
-    file_id = _pdf_filename_part(c.get("slug"), c.get("id"), slug_or_id, fallback="f1")
-    return _pdf_response(
-        pdf_certificates(certificates, pdf_sponsors=sponsors, pdf_branding=branding),
-        f"urkunden_f1_championship_{file_id}.pdf",
-    )
-
-
-@pdf_router.get("/f1/{slug_or_id}/championship-certificates/{user_id}.pdf")
-async def pdf_f1_championship_certificate(slug_or_id: str, user_id: str, access: Optional[str] = None, user: dict | None = Depends(get_optional_user)):
-    db = get_db()
-    from routes.f1_routes import _get_visible_challenge, championship_standings as f1_champ
-    c = await _get_visible_challenge(slug_or_id, user, access=access)
-    if not await _result_export_allowed(c, user):
-        raise HTTPException(status_code=403, detail="Urkunden sind erst nach Challenge-Ende öffentlich.")
-    cs = await f1_champ(c["id"], access, user)
-    rows = cs.get("standings") or []
-    row = next((item for item in rows if item.get("user_id") == user_id), None)
-    if not row:
-        raise HTTPException(status_code=404, detail="Platzierung nicht gefunden.")
-    if row not in _top_certificate_rows(rows):
-        raise HTTPException(status_code=404, detail="Urkunden werden für die ersten vier Plätze erzeugt.")
-    sponsors = await _pdf_sponsors(db)
-    branding = await _pdf_branding(db)
-    file_id = _pdf_filename_part(c.get("slug"), c.get("id"), slug_or_id, fallback="f1")
-    player_id = _pdf_filename_part(row.get("display_name"), user_id, fallback="fahrer")
-    return _pdf_response(
-        pdf_certificate(
-            _certificate_source_from_item(c, "Fast Lap Championship"),
-            row,
-            category="Gesamtwertung",
-            metrics=_f1_championship_certificate_metrics(row),
-            pdf_sponsors=sponsors,
-            pdf_branding=branding,
-        ),
-        f"urkunde_f1_championship_{file_id}_{player_id}.pdf",
-    )
-
-
-# ---------- Audit ----------
-audit_router = APIRouter(prefix="/api/audit", tags=["audit"])
-
-
-@audit_router.get("")
-async def list_audit(action: Optional[str] = None, limit: int = 200, me: dict = Depends(require_club_admin())):
-    db = get_db()
-    q = {}
-    if action:
-        q["action"] = {"$regex": _safe_regex(action)}
-    safe_limit = max(1, min(int(limit or 200), 500))
-    logs = await db.audit_logs.find(q, {"_id": 0}).sort("created_at", -1).to_list(safe_limit)
-    # Enrich actor
-    ids = list({l.get("actor_id") for l in logs if l.get("actor_id")})
-    users = {u["id"]: u for u in await db.users.find({"id": {"$in": ids}},
-                                                      {"_id": 0, "password_hash": 0, "mfa_secret": 0, "mfa_pending_secret": 0, "mfa_recovery_code_hashes": 0}).to_list(500)}
-    for l in logs:
-        if l.get("actor_id"):
-            u = users.get(l["actor_id"], {})
-            l["actor_username"] = u.get("username")
-            l["actor_display_name"] = u.get("display_name")
-    return logs
