+++ backend/tests/test_competition_consumer_boundary_unit.py
@@ -7,7 +7,12 @@ READ_CONSUMERS = (
     "badges.py",
     "pdf_service.py",
     "routes/event_routes.py",
-    "routes/extras_routes.py",
+    "routes/audit_routes.py",
+    "routes/dsgvo_routes.py",
+    "routes/export_routes.py",
+    "routes/season_routes.py",
+    "routes/settings_routes.py",
+    "routes/widget_routes.py",
     "routes/station_routes.py",
     "services/match_notifications.py",
     "services/match_overview.py",
